% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddControlData.ProcessReptMeas.R
\name{AddControlData.ProcessReptMeas}
\alias{AddControlData.ProcessReptMeas}
\title{AddControlData.ProcessReptMeasChngBaseline}
\usage{
\method{AddControlData}{ProcessReptMeas}(cISAData, cSourceData, nISA)
}
\arguments{
\item{cISAData}{ISA data object}

\item{cSourceData}{Source data object containing control data}

\item{nISA}{ISA number}
}
\description{
{ Control data when processed as a Repeated measure change from outcome.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/AddControlData.ProcessReptMeas.R}{View Code on GitHub} }
}
