% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddControlData.R
\name{AddControlData}
\alias{AddControlData}
\title{AddControlData}
\usage{
AddControlData(cISAData, cSourceData, nISA)
}
\arguments{
\item{cISAData}{The ISA object, the class( cISAData ) will determine which version of AddControlData is utilized.}

\item{cSourceData}{The object with the patient data to be used as the patient database for pulling control data from.  THis object must have
the following elements:
1) cSourceData$vIND
2) cSourceData$vOut
3) cSourceData$vTime
4) cSourceData$vTrt}

\item{nISA}{An integer for the ISA number.  As the data is added to the control in may be necessary to know what ISA the data came from.}
}
\description{
{Used to include control data.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/AddControlData.R}{View Code on GitHub} }
}
