% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AddPatient}
\alias{AddPatient}
\title{AddPatient}
\usage{
AddPatient(
  lPatOut,
  dCurrentTime,
  nISA,
  nTrt,
  dfCov,
  cEnrolledPats,
  nPrintDetail = 0
)
}
\arguments{
\item{lPatOut}{List of patient outcomes}

\item{dCurrentTime}{Current simulation time}

\item{nISA}{ISA number}

\item{nTrt}{Treatment number}

\item{dfCov}{Data frame of covariates}

\item{cEnrolledPats}{Enrolled patients data structure}

\item{nPrintDetail}{Print detail level (default 0)}
}
\description{
{This function uses the lPatOut as a database of patients, it pulls 1 patient out and adds it to lEnrolledPats}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/Functions.R}{View Code on GitHub} }
}
