% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AddPatint}
\alias{AddPatint}
\alias{AddPatient}
\title{AddPatient}
\usage{
AddPatient(lPatOut, dCurrentTime, nISA, nTrt, dfCov, cEnrolledPats,
  nPrintDetail = 0)
}
\description{
{This function uses the lPatOut as a database of patients, it pulls 1 patient out and adds it to lEnrolledPats}
}
