% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{all_equal}
\alias{all_equal}
\title{Determine whether two objects are equal.}
\usage{
all_equal(x, y)
}
\arguments{
\item{x}{An object.}

\item{y}{Another object.}
}
\value{
Logical.
}
\description{
Determine whether two objects are equal.
}
\examples{
all_equal(data.frame(y = 1:2), data.frame(y = c(2, 4)))
}
