% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{auto.library}
\alias{auto.library}
\title{Load package, install automatically if missing.}
\usage{
auto.library(package)
}
\arguments{
\item{package}{the name of a package}
}
\value{
invisible \code{TRUE}
}
\description{
Load \code{package}, install highest version
from \href{https://www.bioconductor.org/install/}{bioconductor}
or \href{http://cran.rstudio.com/}{rstudio} if missing.
}
\examples{
{
    auto.library(stringr) #use package name
    auto.library("stringr") #use package name string
    auto.library("package_no_exist") # Print warning message when package is not found



tutorial("auto.library") # print tutorial to console

}

