% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{autosnake}
\alias{autosnake}
\title{Convert dataframe names to snake_case.}
\usage{
autosnake(df, ...)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
replacements Named vector of replacements.
}
\description{
Convert dataframe names to snake_case.
}
\examples{
library(magrittr)
data.frame("Speed km/hr" = 1:5, check.names = FALSE) \%>\% autosnake(c("/" = "per"))
}
