% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildResultsDataSet.R
\name{BuildResultsDataSet}
\alias{BuildResultsDataSet}
\title{BuildResultsDataSet}
\usage{
BuildResultsDataSet(strMainResultsFile, vStrISAResultsFile, vISANumber)
}
\arguments{
\item{strMainResultsFile}{The main output file produced when running a simulation.  Typically, it is found in the out/ subdirectory.}

\item{vStrISAResultsFile}{The file that contains the output from in ISA.  For ISA XXX it is typically the ISAXXX/ subdirectory.}

\item{vISANumber}{{An integer that will be used to prefix all of the ISA output variables names.  For example, if nISANumber =1, then all variables names
in the file will be prefixed with ISA1.}}
}
\description{
{This function is used to read in the input files and created a combined output file.    }
}
