% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSimulationResultsDataSet.R
\name{BuildSimulationResultsDataSet}
\alias{BuildSimulationResultsDataSet}
\title{BuildSimulationResultsDataSet}
\usage{
BuildSimulationResultsDataSet()
}
\description{
{This function is used to read the files in the out and ISAOut1, ISAOut2,..., directories and combine everything into one simulation dataset
that can be used for creating simulation results.  You should set the R working directory that contains the main, ISAOut1, ISOut2,...., directories.
This function is intended to streamline and replace the BuildResultsDataSet funciton.  }
}
