% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CheckGlobalVariables}
\alias{CheckGlobalVariables}
\title{CheckGlobalVariables}
\usage{
CheckGlobalVariables()
}
\description{
{Determine if the global variables that are needed have been defined and if not define them. Use of global variables
should be restricted to the variables that are used throughout the code and only in VERY limited use.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/Functions.R}{View Code on GitHub} }
}
