% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckISAEnrollmentStatus.R
\name{CheckISAEnrollmentStatus}
\alias{CheckISAEnrollmentStatus}
\title{CheckISAEnrollmentStatus}
\usage{
CheckISAEnrollmentStatus(cRandomizer, vISAStatus, dfCov)
}
\arguments{
\item{cRandomizer}{Randomizer object}

\item{vISAStatus}{Vector of ISA status values}

\item{dfCov}{Data frame of patient covariates}
}
\description{
CheckISAEnrollmentStatus{ }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CheckISAEnrollmentStatus.R}{View Code on GitHub} }
}
