% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckNextTrialAnalysisTime.R
\name{CheckNextTrialAnalysisTime}
\alias{CheckNextTrialAnalysisTime}
\title{CheckNextTrialAnalysisTime}
\usage{
CheckNextTrialAnalysisTime(
  cISADesigns,
  lEnrolledPats,
  vISAStatus,
  dCurrentTime,
  vISAAnalysisIndx,
  vPreviousIATime
)
}
\arguments{
\item{cISADesigns}{ISA designs object}

\item{lEnrolledPats}{List of enrolled patients}

\item{vISAStatus}{Vector of ISA status values}

\item{dCurrentTime}{Current simulation time}

\item{vISAAnalysisIndx}{Vector of ISA analysis indices}

\item{vPreviousIATime}{Vector of previous interim analysis times}
}
\description{
{This function will Check the monitoring rule and return the time of next Analysis for the trial}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CheckNextTrialAnalysisTime.R}{View Code on GitHub} }
}
