% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePosteriorProbs.R
\name{ComputePosteriorProbs.default}
\alias{ComputePosteriorProbs.default}
\title{ComputePosteriorProbs.default}
\usage{
\method{ComputePosteriorProbs}{default}(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis, lSamples)
}
\arguments{
\item{nISAAnalysisIndx}{A the index of the analysis}

\item{bIsFinalISAAnalysis}{TRUE or FALSE to indicate if this is the final analysis. Typically, used
in a Bayesian design such that the final analysis can have different cutoff values.}

\item{lSamples}{List with two vectors, vPostSampPlac and vPostSampTrt, that are samples of the posterior of
\eqn{\theta} for the placebo and treatment, respectively.}

\item{cAnalysis$vMAVCutoff}{A vector of cutoffs for the MAV at each analysis.}

\item{cAnalysis$vTVCutoff}{A vector of cutoffs for the TV.}
}
\value{
List with two valued dMAVCutoff and dTVCutoff

List with four values dPrGrtMAV, dPrGrtTV, dMAVCutoff and dTVCutoff
}
\description{
{Used in a trial with a Bayesian analysis and parameter of interest \eqn{\theta}, where
the decisions are based two posterior probabilities Pr( \eqn{\theta} > MAV |data ) > vMAVCutoff
and Pr( \eqn{\theta} > TV | data ) > vTVCutoff.
This function returns a list with dMAVCutoff and dTVCutoff.  0 <= dMAVCutoff and dTVCutoff <= 1}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/ComputePosteriorProbs.R}{View Code on GitHub} }
}
