% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePosteriorProbs.R
\name{ComputePosteriorProbs.MAVOnly}
\alias{ComputePosteriorProbs.MAVOnly}
\title{ComputePosteriorProbs.MAVOnly}
\usage{
\method{ComputePosteriorProbs}{MAVOnly}(
  cAnalysis,
  nISAAnalysisIndx,
  bIsFinalISAAnalysis,
  lSamples
)
}
\arguments{
\item{nISAAnalysisIndx}{A the index of the analysis}

\item{bIsFinalISAAnalysis}{TRUE or FALSE to indicate if this is the final analysis. Typically, used
in a Bayesian design such that the final analysis can have different cutoff values.
dTVCutoff}

\item{lSamples}{List with two vectors, vPostSampPlac and vPostSampTrt, that are samples of the posterior of
\eqn{\theta} for the placebo and treatment, respectively.}

\item{cAnalysis$vMAVCutoff}{A vector of cutoffs for the MAV at each analysis.}

\item{cAnalysis$vTVCutoff}{A vector of cutoffs for the TV.}
}
\value{
List with three values dPrGrtMAV,  dMAVCutoff and dTVCutoff
}
\description{
{Used in a trial with a Bayesian analysis and parameter of interest \eqn{\theta}, where
the decisions are based one posterior probabilities Pr( \eqn{\theta_P} - \eqn{\theta_T} > MAV |data ).
This function returns a list with dPrGrtMAV, dMAVCutoff and dTVCutoff.  0 <= dMAVCutoff and dTVCutoff <= 1}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/ComputePosteriorProbs.R}{View Code on GitHub} }
}
