% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePosteriorProbs.R
\name{ComputePosteriorProbs.MAVTarget}
\alias{ComputePosteriorProbs.MAVTarget}
\title{ComputePosteriorProbs.MAVTarget}
\usage{
\method{ComputePosteriorProbs}{MAVTarget}(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis, lSamples)
}
\arguments{
\item{nISAAnalysisIndx}{A the index of the analysis}

\item{bIsFinalISAAnalysis}{TRUE or FALSE to indicate if this is the final analysis. Typically, used
in a Bayesian design such that the final analysis can have different cutoff values.
dTVCutoff}

\item{lSamples}{List with vPostSamp that are samples of the posterior of
\eqn{\theta} the parameter of interest.}

\item{cAnalysis$vMAVCutoff}{A vector of cutoffs for the MAV at each analysis.}

\item{cAnalysis$vTVCutoff}{A vector of cutoffs for the TV.}
}
\value{
List with three values dPrGrtMAV,  dMAVCutoff and dTVCutoff
}
\description{
{Used in a trial with a Bayesian analysis and parameter of interest \eqn{\theta}, where
the decisions are based one posterior probabilities Pr(  \eqn{\theta_T} > MAV |data ).
This function returns a list with dPrGrtMAV, dMAVCutoff and dTVCutoff.  0 <= dMAVCutoff and dTVCutoff <= 1}
}
