% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePosteriorProbs.R
\name{ComputePosteriorProbs}
\alias{ComputePosteriorProbs}
\title{ComputePosteriorProbs}
\usage{
ComputePosteriorProbs(cAnalysis, nISAAnalysisIndx, bIsFinalISAAnalysis,
  lSamples)
}
\description{
ComputePosteriorProbs
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/ComputePosteriorProbs.R}{View Code on GitHub} }
}
