% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyControlCovariates.R
\name{CopyControlCovariates}
\alias{CopyControlCovariates}
\title{CopyControlCovariates}
\usage{
CopyControlCovariates(cDestData, cSourceData, vSubsetCtrl)
}
\description{
{Used to include control data with covaraites}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CopyControlCovariates.R}{View Code on GitHub} }
}
