% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariates.R
\name{CopyCovariates.default}
\alias{CopyCovariates.default}
\title{CopyCovariates.default}
\usage{
\method{CopyCovariates}{default}(lSimDest, lSimSource)
}
\description{
This function will copy elements  lSimSource$vCov1,..., lSimSource$vCovXX to lSimDet and return the updated lSimDet.
  The names vCov1, vCov2,... must be consecutive and cannot skip eg, vCov1, vCov3 will the function will stop after copying vCov1.
}
