% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create.info}
\alias{create.info}
\title{Print info with timestamp}
\usage{
create.info(prefix = "")
}
\arguments{
\item{prefix}{prefix of message.}
}
\value{
object of class \code{info}
}
\description{
Print debug message, progress reports... with timestamp.
\itemize{
  \item Default on
  \item Turn off by \code{info.enable(FALSE)}
  \item Turn on by \code{info.enable(TRUE)}
}
}
\examples{
{
    info.self <- create.info() #create object of info class
    print(info.self,"print message without prefix.")

    info.self <- create.info("tutorial.info") #create object of info class with "tutorial.info" as prefix
    print(info.self,"print message with prefix, use identifier for prefix usually.")

    # turn it off
    info.enable(FALSE) #turn information printing off all together
    print(info.self,"will print nothing from now on.")
    print(info.self,"conform print nothing.")

    # turn it on
    info.enable(TRUE) #turn information printing on all together
    print(info.self,"start printing again.")
  }



tutorial("info") # print tutorial to console

}

