% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataSetForAnalysis.R
\name{CreateDataSetForAnalysis}
\alias{CreateDataSetForAnalysis}
\title{CreateDataSetForAnalysis}
\usage{
CreateDataSetForAnalysis(cPats, dCurrentTime, vISAStatus)
}
\description{
{This function is used to created the dataset for analysis.  }
}
