% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProject.R
\name{CreateProject}
\alias{CreateProject}
\title{CreateProject}
\usage{
CreateProject(
  strProjectDirectory = "",
  strProjectName = "NewProject",
  strAnalysisName = "NewAnalysisName",
  strSimPatientOutcomeName = "NewSimOutcomeName",
  strBorrowing = "NoBorrowing",
  nQtyReps = 10,
  mQtyPatientsPerArm = NULL,
  dQtyMonthsFU = 1,
  vISAStartTimes = NULL,
  bCreateProjectSubdirectory = TRUE
)
}
\arguments{
\item{strProjectDirectory}{The complete path to where the project is created.  If a complete path
is not provided then the directory is relative to the working directory.  If this is left blank, the current working directory is used.}

\item{strProjectName}{The name of the project.   A directory with the name strProjectName is created in strProjectDirectory if bCreateProjectSubdirectory = TRUE}

\item{strAnalysisName}{The name of the analysis that should be done. If the analysis is part of OCTOPUS, then no new
files are created.  If the analysis is new, a file named RunAnalysis.strAnalysisName.R is created.}

\item{strSimPatientOutcomeName}{The name of the patient simulator that should be utilized. If the patient simulator
is part of OCTOPUS, then no new
files are created.  If the patient simulator is new, a file named SimPatientOutocmes.strSimPatientOutcomeName.R is created.}

\item{strBorrowing}{Specify the option for sharing control patients across ISAs.  If strBorrowing = "NoBorrowing"
then patients are not shared between ISAs.  If strBorrowing = "AllControls" then all control patients are included in
the data set that is sent to the RunAnalysis funciton.}

\item{nQtyReps}{The number of replications to simulate for each scenario.}

\item{mQtyPatientsPerArm}{Matrix with 2 or more columns to define the number of patients per arm on each ISA.  Each row represents an ISA and
column 1 the number of patients on control/placebo and column 2,... the number of patients on the treatment arms in the ISA.  Each ISA must have
the same number of arms (a control or placbo and treatment arms(s) ).  This option can be chnaged once the project is created by modifing the
SetupTrialDesign file.  If this parameter is NULL you will need to change the options in the BuildMe.R file that is created as part of the project.}

\item{dQtyMonthsFU}{The number of months of follow-up the last patient before running the final analysis (FA). By default, the design created
only includes a final analysis with NO interim analysis.  To use another option please see the TrialDesign.R file in the projected that is created.}

\item{bCreateProjectSubdirectory}{If TRUE the a subdirectory with the project name will be created, if FALSE then the project is created in strProjectDirectory.}
}
\description{
Created an R Studio project with example files to simulate a platform trial utilizing the OCTOPUS package.
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CreateProject.R}{View Code on GitHub} }
}
