% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_connection.R
\name{delete_table_mysql}
\alias{delete_table_mysql}
\title{A Database Specific Function for Dropping Tables}
\usage{
delete_table_mysql(con, schema, table)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A string containing the schema name.}

\item{table}{A string containing the table name.}
}
\value{
A result string. Either "Success" or an error message.
}
\description{
A Database Specific Function for Dropping Tables
}
\keyword{internal}
