% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{despace}
\alias{despace}
\title{Function to remove factors and spaces and NA from dataframe prior to writing.}
\usage{
despace(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Function to remove factors and spaces and NA from dataframe prior to writing.
}
\examples{
data.frame(x = c(NA, "A string"), y = 1:2) \%>\% despace()
}
