% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dotty}
\alias{dotty}
\title{Visualise dataframe contents using ggplot.}
\usage{
dotty(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A ggplot object.
}
\description{
Visualise dataframe contents using ggplot.
}
\examples{
library(magrittr)
data.frame(A = -10:10, B = NA, C = "C", stringsAsFactors = FALSE) \%>\% dotty() \%>\% print()
}
