% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ensnakeify}
\alias{ensnakeify}
\title{Convert strings to snake_case.}
\usage{
ensnakeify(x, replacements = c(`\%` = "pc"))
}
\arguments{
\item{x}{A vector of strings.}
}
\value{
replacements Named vector of replacements.
}
\description{
Convert strings to snake_case.
}
\examples{
ensnakeify(c("Speed km/hr"), c("/" = "per"))
}
