% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connection.R
\name{establish_connection_postgres}
\alias{establish_connection_postgres}
\title{Create A Postgres Database Connection}
\usage{
establish_connection_postgres(database_credentials)
}
\arguments{
\item{database_credentials}{A data frame with connection specific credentials.}
}
\value{
A database connection object.
}
\description{
Create A Postgres Database Connection
}
