% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNumberOfPatientsToSimulate.R
\name{EstimateNumberOfPatientToSimulate.Categorical}
\alias{EstimateNumberOfPatientToSimulate.Categorical}
\alias{EstimateNumberOfPatientsToSimulate.Categorical}
\title{EstimateNumberOfPatientToSimulate.Categorical}
\usage{
\method{EstimateNumberOfPatientsToSimulate}{Categorical}(cSimCovariates,
  nMaxQtyPats)
}
\description{
EstimateNumberOfPatientToSimulate.Categorical {returns nMaxQtyPats/min(vProbs)}
}
