% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNumberOfPatientsToSimulate.R
\name{EstimateNumberOfPatientToSimulate.default}
\alias{EstimateNumberOfPatientToSimulate.default}
\alias{EstimateNumberOfPatientsToSimulate.default}
\title{EstimateNumberOfPatientToSimulate.default}
\usage{
\method{EstimateNumberOfPatientsToSimulate}{default}(cSimCovariates,
  nMaxQtyPats)
}
\description{
EstimateNumberOfPatientsToSimulate.default{ Call the EstimateNumberOfPatientsToSimulate for each covaiate and returns the maximum of the results.}
}
