% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNumberOfPatientsToSimulate.R
\name{EstimateNumberOfPatientToSimulate.NULL}
\alias{EstimateNumberOfPatientToSimulate.NULL}
\alias{EstimateNumberOfPatientsToSimulate.NULL}
\title{EstimateNumberOfPatientToSimulate.NULL}
\usage{
\method{EstimateNumberOfPatientsToSimulate}{NULL}(cSimCovariates,
  nMaxQtyPats)
}
\description{
EstimateNumberOfPatientToSimulate.NULL {This opion is used when no covariates are in a design. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/EstimateNumberOfPatientsToSimulate.R}{View Code on Gitub} }
}
