% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNumberOfPatientsToSimulate.R
\name{EstimateNumberOfPatientToSimulate}
\alias{EstimateNumberOfPatientToSimulate}
\alias{EstimateNumberOfPatientsToSimulate}
\title{EstimateNumberOfPatientToSimulate}
\usage{
EstimateNumberOfPatientsToSimulate(cSimCovariates, nMaxQtyPats)
}
\description{
EstimateNumberOfPatientToSimulate {}
}
