% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{excel_to_date}
\alias{excel_to_date}
\title{Convert Excel date numbers to Date.}
\usage{
excel_to_date(excel)
}
\arguments{
\item{excel}{A vector of excel date numbers.}
}
\value{
A vector of dates.
}
\description{
Convert Excel date numbers to Date.
}
\examples{
excel_to_date(40000)
}
