% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fillgaps}
\alias{fillgaps}
\title{Function to fill gaps with first available value.}
\usage{
fillgaps(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector.
}
\description{
Function to fill gaps with first available value.
}
\examples{
data.frame(x = c(1, 1, NA, 1, 1)) \%>\% dplyr::mutate(x = fillgaps(x))
}
