% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connections.R
\name{get_connections}
\alias{get_connections}
\title{Get existing saved database connections.}
\usage{
get_connections()
}
\value{
A named character vector of database connection names and ids.
}
\description{
Get existing saved database connections.
}
