% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_connection.R
\name{get_n_rows_mysql}
\alias{get_n_rows_mysql}
\title{A Database Specific Function to Retrieve the Number of Rows of a Table}
\usage{
get_n_rows_mysql(con, schema, table, query = "")
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A string containing the schema name.}

\item{table}{A string containing the table name.}

\item{query}{A string containing the query to send.}
}
\value{
An integer for the number of rows in the table.
}
\description{
A Database Specific Function to Retrieve the Number of Rows of a Table
}
