% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connection.R
\name{get_preview_postgres}
\alias{get_preview_postgres}
\title{A Database Specific Function for Retrieving Table Previews}
\usage{
get_preview_postgres(con, shema, table)
}
\arguments{
\item{con}{A database connection object.}

\item{table}{A string containing the table name.}

\item{schema}{A string containing the schema name.}
}
\value{
A data frame of 100 rows from the database table.
}
\description{
A Database Specific Function for Retrieving Table Previews
}
