% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake_connection.R
\name{get_preview_snowflake}
\alias{get_preview_snowflake}
\title{A Database Specific Function for Retrieving Table Previews}
\usage{
get_preview_snowflake(con, schema, table)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A string containing the schema name.}

\item{table}{A string containing the table name.}
}
\value{
A data frame of 100 rows from the database table.
}
\description{
A Database Specific Function for Retrieving Table Previews
}
\keyword{internal}
