% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connection.R
\name{get_schemas_postgres}
\alias{get_schemas_postgres}
\title{A Database Specific Function To Retrieve Database Schemas}
\usage{
get_schemas_postgres(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A character vector of all schemas in the database.
}
\description{
A Database Specific Function To Retrieve Database Schemas
}
\keyword{internal}
