% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_connection.R
\name{get_tables_mysql}
\alias{get_tables_mysql}
\title{A Database Specific Function to Retrieve All Database Tables}
\usage{
get_tables_mysql(con, schema)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A}
}
\value{
A character vector of all tables in a given schema.
}
\description{
A Database Specific Function to Retrieve All Database Tables
}
