% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake_connection.R
\name{get_tables_snowflake}
\alias{get_tables_snowflake}
\title{A Database Specific Function to Retrieve All Database Tables}
\usage{
get_tables_snowflake(con, schema)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A}
}
\value{
A character vector of all tables in a given schema.
}
\description{
A Database Specific Function to Retrieve All Database Tables
}
\keyword{internal}
