% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBayesianCutoffs.R
\name{GetBayesianCutoffs.default}
\alias{GetBayesianCutoffs.default}
\title{GetBayesianCutoffs.default}
\usage{
\method{GetBayesianCutoffs}{default}(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\arguments{
\item{nISAAnalysisIndx}{An integer with the index of the analysis}

\item{bIsFinalISAAnalysis}{TRUE or FALSE to indicate if this is the final analysis. Typically, used
in a Bayesian design such that the final analysis can have different cutoff values.}

\item{cAnalysis$vMAVCutoff}{A vector of cutoffs for the MAV at each analysis.}

\item{cAnalysis$vTVCutoff}{A vector of cutoffs for the TV.}
}
\value{
List with two values dMAVCutoff and dTVCutoff.
}
\description{
{Used in a trial with a Bayesian analysis and parameter of interest \eqn{\theta}, where
the decisions are based two posterior probabilities Pr( \eqn{\theta} > MAV |data ) > vMAVCutoff
and Pr( \eqn{\theta} > TV | data ) > vTVCutoff.
This function returns a list with dMAVCutoff and dTVCutoff.  0 <= dMAVCutoff and dTVCutoff <= 1.
Typically, this function is not called directly but as part of the ComputePosteriorProb function.}
}
