% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBayesianCutoffs.R
\name{GetBayesianCutoffs.MAVOnly}
\alias{GetBayesianCutoffs.MAVOnly}
\title{GetBayesianCutoffs.MAVOnly}
\usage{
\method{GetBayesianCutoffs}{MAVOnly}(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\arguments{
\item{cAnalysis$vPUpper}{Vector of upper cutoffs, typically used in the context of success if a posterior probability > vPUpper,
0 <=  vPLowerCutoff <= vPUpperCutoff <= 1}

\item{cAnalysis$vPLower}{Vector of lower cutoffs, typically used in the context of failure if a posterior probability < vPLower,
0 <=  vPLowerCutoff <= vPUpperCutoff <= 1}

\item{cAnalysis$dFinalPUpper}{Value of upper cutoff used at the FINAL analysis, typically used in the context of success if a posterior probability > dFinalPUpper,
0 <= dPLowerCutoff <= dPUpperCutoff <=  1}

\item{cAnalysis$dFinalPLower}{Value of lower cutoff used at the FINAL analysis, typically used in the context of failure if a posterior probability < dFinalPLower,
0 <= dPLowerCutoff <= dPUpperCutoff <=  1}
}
\description{
{Used in a Bayesian analysis when you have interest in determining if a single posterior probability, eg probability a
parameter is greater than MAV,
being greater than an upper cutoff or below a lower cutoff.  In particular, you must
define vPUpper, vPLower, dFinalPUpper, dFinalPLower as part of your cAnalysis object.
This function returns a list with dPUpperCutoff and dPLowerCutoff.  0 <= dPUpperCutoff <= dPLowerCutoff <= 1}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/GetBayesianCutoffs.R}{View Code on GitHub} }
}
