% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBayesianCutoffs.R
\name{GetBayesianCutoffs.MAVTarget}
\alias{GetBayesianCutoffs.MAVTarget}
\title{GetBayesianCutoffs.MAVTarget}
\usage{
\method{GetBayesianCutoffs}{MAVTarget}(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\arguments{
\item{cAnalysis$vPUpper}{Vector of upper cutoffs, typically used in the context of success if a posterior probability > vPUpper,
0 <=  vPLowerCutoff <= vPUpperCutoff <= 1}

\item{cAnalysis$vPLower}{Vector of lower cutoffs, typically used in the context of failure if a posterior probability < vPLower,
0 <=  vPLowerCutoff <= vPUpperCutoff <= 1}

\item{cAnalysis$dFinalPUpper}{Value of upper cutoff used at the FINAL analysis, typically used in the context of success if a posterior probability > dFinalPUpper,
0 <= dPLowerCutoff <= dPUpperCutoff <=  1}

\item{cAnalysis$dFinalPLower}{Value of lower cutoff used at the FINAL analysis, typically used in the context of failure if a posterior probability < dFinalPLower,
0 <= dPLowerCutoff <= dPUpperCutoff <=  1}
}
\description{
{This function is the same as GetBayesianCutoffs.MAVOnly as MAVTarget inheirts from MAVOnly }
}
