% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFinalISAAnalysisTime.R
\name{GetFinalISAAnalysisTime.default}
\alias{GetFinalISAAnalysisTime.default}
\title{#' export
GetFinalISAAnalysisTime.default <- function( cISADesigns, nISA, lEnrolledPats )
{
    stop( paste("ERROR: The GetFinalISAAnalysisTime.default is not defined class( GetFinalISAAnalysisTime )=", class(cISADesigns) ) )
}}
\usage{
\method{GetFinalISAAnalysisTime}{default}(cISADesigns, nISA, lEnrolledPats)
}
\description{
#' export
GetFinalISAAnalysisTime.default <- function( cISADesigns, nISA, lEnrolledPats )
{
    stop( paste("ERROR: The GetFinalISAAnalysisTime.default is not defined class( GetFinalISAAnalysisTime )=", class(cISADesigns) ) )
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/GetFinalISAAnalysisTime.R}{View Code on GitHub} }
}
