% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFinalISAAnalysisTime.R
\name{GetFinalISAAnalysisTime}
\alias{GetFinalISAAnalysisTime}
\title{GetFinalISAAnalysisTime}
\usage{
GetFinalISAAnalysisTime(cISADesigns, nISA, lEnrolledPats)
}
\description{
{ Compute the final analysis time for a given ISA based on the current data}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/GetFinalISAAnalysisTime.R}{View Code on GitHub} }
}
