% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.Functions.R
\name{GetLSDiffGLS}
\alias{GetLSDiffGLS}
\title{GetLSDiffGLS}
\usage{
GetLSDiffGLS(glsFit, nTrt, nTime, dMAV, dTV, dLowerCI, dUpperCI, bPlacMinusTrt)
}
\arguments{
\item{glsFit:}{The results of the lme() model fit, required to have the variable TRT and Time}

\item{nTrt:}{Integer value for the Treatment to compare to 0 (placebo)}

\item{nTime:}{Integer value for the time variable that should be use.  For example 24, for 24}
}
\description{
{This function will produce the estiamtes comparing to treatment 0 for the LME models that are fit.
  The Z values will match to SAS, which is what is used in decision making.
  This version requires glsFit requires glsFit to be a fit for gls}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunAnalysis.Functions.R}{View Code on GitHub} }
}
