% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentComparitor.R
\name{ggPlotPlatformRecruitment}
\alias{ggPlotPlatformRecruitment}
\title{PlotPlatformRecruitment}
\usage{
ggPlotPlatformRecruitment(nQtyReps, nMaxQtyPats, vPatsPerMonthPerSite1,
  dDelayToStartPOC, dDelayBetweenTrialsPOC, vQtySitesPOC,
  dDelayToStartPlat, dDelayBetweenTrialsPlat, vQtyOfSitesPlat1,
  vQtyOfSitesPlat2)
}
\description{
{This function creates a plot of recruitment comparing a platform with 2 ISAs and 2 consecutive POC studies.
Mainly used by the Shiny app execute by calling RunExample( "CompareRecruitment" ).  This version returns a ggplot2 object. }
}
