% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeISARandomizer.R
\name{InitializeISARandomizer.default}
\alias{InitializeISARandomizer.default}
\title{InitializeISARandomizer.default}
\usage{
\method{InitializeISARandomizer}{default}(cISARandomizer, dISAStartTime)
}
\description{
{This function is used to initialize the randomzier for an ISA. No default option
is specified to make sure the user selects a valid randomizer. }
}
