% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeISARandomizer.R
\name{InitializeISARandomizer.POCRandomizer}
\alias{InitializeISARandomizer.POCRandomizer}
\title{InitializeISARandomizer.POCRandomizer}
\usage{
\method{InitializeISARandomizer}{POCRandomizer}(cISARandomizer, dISAStartTime)
}
\arguments{
\item{cISARandomizer}{The cISARandomizer$vQtyPatsInit identifies the initial patients to randomize first, the vQtyPatsInit is part of the vQtyPats for that ISA
and is not required by other randomizer so this function will stop if it is not defined}

\item{dISAStartTime}{Parameter is not utilized for POCRandomizer}
}
\description{
{ InitializeISARandomizer.POCRandomizer will randomize the first patients to a select set of
arms or doses in the ISA (eg a POC phase).  The remaining arms or doses are opened based on the number of initial patients.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/InitializeISARandomizer.R}{View Code on GitHub} }
}
