% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeISARandomizer.R
\name{InitializeISARandomizer}
\alias{InitializeISARandomizer}
\title{InitializeISARandomizer}
\usage{
InitializeISARandomizer(cISARandomizer, dISAStartTime)
}
\description{
{This function is used to initialize the randomzier for an ISA. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/InitializeISARandomizer.R}{View Code on GitHub} }
}
