% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeTrialRandomizer.R
\name{InitializeTrialRandomizer.default}
\alias{InitializeTrialRandomizer.default}
\alias{InitializeTrialRandomizer}
\title{InitializeTrialRandomizer.default}
\usage{
InitializeTrialRandomizer(cTrialDesign, vISAStartTimes)
}
\arguments{
\item{cTrialDesign}{Trial design opbject giving specifics about the trial design.}

\item{vISAStartTime}{A vector of times that each ISA opens/starts in the trial.  A vector of c(0,...) would indicate all ISAs start at the start of the trial.}
}
\description{
{This function is used to initialize the trial randomizer. The default version each
ISA and calls the InitializeISARandomizer.  Since each ISA can have a specific randomizer  the default
for the TrialRandomizer is just to initialize each ISA randomizer.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/InitializeTrialRandomizer.R}{View Code on GitHub} }
}
