% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_loaded}
\alias{is_loaded}
\title{Check whether a package is loaded.}
\usage{
is_loaded(package_name)
}
\arguments{
\item{package_name}{A string.}
}
\value{
Logical.
}
\description{
Check whether a package is loaded.
}
\examples{
is_loaded("dplyr")
}
