% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_drivers.R
\name{list_drivers}
\alias{list_drivers}
\title{List Compatible Database Drivers}
\usage{
list_drivers()
}
\value{
A character vector of compatible database drivers.
}
\description{
List Compatible Database Drivers
}
