% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.General}
\alias{MakeDecision.General}
\title{MakeDecision.General}
\usage{
\method{MakeDecision}{General}(lDecision, lResAnalysis, bFinalAnalysis,
  cRandomizer)
}
\description{
{This is the General version of Make decisions that allows for one approach at IAs (IA)
and another (which could be the same) at the Final Analysis (FA).  You must specify
lDecision$strApproachIA, method used during the IAs, and lDecision$strApproachFA, method used at the FA.
The options for strApproachFA and strApproachIA are any of the MakeDecision options like MakeDecision.TwoOutcomesOption1 or
MakeDecision.TwoOutcomeOptions2.
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecision.R}{View Code on GitHub} }
}
