% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.Outcome1Only}
\alias{MakeDecision.Outcome1Only}
\title{MakeDecision.Outcome1Only}
\usage{
\method{MakeDecision}{Outcome1Only}(lDecision, lResAnalysis,
  bFinalAnalysis, cRandomizer)
}
\description{
{This option ONLY uses the decision from Outcome 1.
\tabular{ccccc}{
            \tab G |  \tab  NG \tab P   \tab G \cr
  Outcome 2 \tab P |  \tab  NG \tab P   \tab G \cr
            \tab NG|  \tab  NG \tab P   \tab G \cr
            \tab ---  \tab---  \tab --- \tab--- \cr
            \tab      \tab  NG \tab P   \tab G \cr
            \tab      \tab     \tab Outcome 1 \tab \cr
}
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecision.R}{View Code on GitHub} }
}
