% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.Outcome2Only}
\alias{MakeDecision.Outcome2Only}
\title{MakeDecision.Outcome2Only}
\usage{
\method{MakeDecision}{Outcome2Only}(lDecision, lResAnalysis,
  bFinalAnalysis)
}
\description{
{This option ONLY uses the decision from Outcome 2.
\tabular{ccccc}{
            \tab G |  \tab  G  \tab G   \tab G \cr
  Outcome 2 \tab P |  \tab  P  \tab P   \tab P \cr
            \tab NG|  \tab  NG \tab NG  \tab NG \cr
            \tab ---  \tab---  \tab --- \tab--- \cr
            \tab      \tab  NG \tab P   \tab G \cr
            \tab      \tab     \tab Outcome 1 \tab \cr
}
}
}
