% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision}
\alias{MakeDecision}
\title{MakeDecision}
\usage{
MakeDecision(lDecision, lResAnalysis, bFinalAnalysis, cRandomizer)
}
\value{
A list like the following:  list( nGo = nGo, nNoGo = nNoGo, nPause = nPause ) with only one of the element = 1.
}
\description{
{This function is a generic function. Several options are available. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecision.R}{View Code on GitHub} }
}
