% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.TwoOutcomeOption1}
\alias{MakeDecision.TwoOutcomeOption1}
\title{MakeDecision.TwoOutcomeOption1}
\usage{
\method{MakeDecision}{TwoOutcomeOption1}(lDecision, lResAnalysis,
  bFinalAnalysis, cRandomizer)
}
\description{
{This two outcome decision options utilizes modified OR between two outcomes.
For each outcome you make a decision Go (G), Pause (P) or No Go (NG) and when the two outcomes are
combined the following tabled describes how decisions will be made.
If either outcome is a G then then the decision is a Go, if it is NOT a Go then if either is a NG
the decision is a NG.  If both are a P then the decision is a P.
\tabular{ccccc}{
            \tab G |  \tab G \tab G  \tab G \cr
  Outcome 2 \tab P | \tab NG \tab P  \tab G \cr
            \tab NG| \tab NG \tab NG \tab G \cr
            \tab --- \tab--- \tab --- \tab--- \cr
            \tab     \tab NG \tab P  \tab G \cr
            \tab     \tab    \tab Outcome 1 \tab \cr
}
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecision.R}{View Code on GitHub} }
}
