% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.TwoOutcomeOption2}
\alias{MakeDecision.TwoOutcomeOption2}
\title{MakeDecision.TwoOutcomeOption2}
\usage{
\method{MakeDecision}{TwoOutcomeOption2}(lDecision, lResAnalysis,
  bFinalAnalysis)
}
\description{
{This two outcome decision options utilizes an AND between two outcomes.
For each outcome you make a decision Go (G), Pause (P) or No Go (NG) and when the two outcomes are
combined the following tabled describes how decisions will be made.
\tabular{ccccc}{
            \tab G |  \tab P \tab P \tab G \cr
  Outcome 2 \tab P | \tab P  \tab P \tab P \cr
            \tab NG| \tab NG \tab P \tab P \cr
            \tab --- \tab--- \tab --- \tab--- \cr
            \tab     \tab NG \tab P \tab G \cr
            \tab     \tab    \tab Outcome 1 \tab \cr
}
}
}
