% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.TwoOutcomeOption3}
\alias{MakeDecision.TwoOutcomeOption3}
\title{MakeDecision.TwoOutcomeOption3}
\usage{
\method{MakeDecision}{TwoOutcomeOption3}(lDecision, lResAnalysis,
  bFinalAnalysis, cRandomizer)
}
\description{
{This two outcome decision options makes if GO if either outcome is a Go and makes a No Go
if BOTH outcomes are a No Go, otherwise a Pause is made.
For each outcome you make a decision Go (G), Pause (P) or No Go (NG) and when the two outcomes are
combined the following tabled describes how decisions will be made.
\tabular{ccccc}{
            \tab G |  \tab G \tab G \tab G \cr
  Outcome 2 \tab P | \tab P  \tab P \tab G \cr
            \tab NG| \tab NG \tab P \tab G \cr
            \tab --- \tab--- \tab --- \tab--- \cr
            \tab     \tab NG \tab P \tab G \cr
            \tab     \tab    \tab Outcome 1 \tab \cr
}
}
}
