% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.TwoOutcomeOption4}
\alias{MakeDecision.TwoOutcomeOption4}
\title{MakeDecision.TwoOutcomeOption4}
\usage{
\method{MakeDecision}{TwoOutcomeOption4}(lDecision, lResAnalysis, bFinalAnalysis, cRandomizer)
}
\description{
{This option is as follows.
\tabular{ccccc}{
            \tab G |  \tab  NG \tab G   \tab G \cr
  Outcome 2 \tab P |  \tab  NG \tab P   \tab G \cr
            \tab NG|  \tab  NG \tab NG  \tab G \cr
            \tab ---  \tab---  \tab --- \tab--- \cr
            \tab      \tab  NG \tab P   \tab G \cr
            \tab      \tab     \tab Outcome 1 \tab \cr
}
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecision.R}{View Code on GitHub} }
}
