% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecision.R
\name{MakeDecision.TwoOutcomeOption7}
\alias{MakeDecision.TwoOutcomeOption7}
\title{MakeDecision.TwoOutcomeOption7}
\usage{
\method{MakeDecision}{TwoOutcomeOption7}(lDecision, lResAnalysis,
  bFinalAnalysis)
}
\description{
{This two outcome decision options priority on a Go for Outcome 1
\tabular{ccccc}{
            \tab G |  \tab  P  \tab G   \tab G \cr
  Outcome 2 \tab P |  \tab  NG \tab P   \tab G \cr
            \tab NG|  \tab  NG \tab P   \tab G \cr
            \tab ---  \tab---  \tab --- \tab--- \cr
            \tab      \tab  NG \tab P   \tab G \cr
            \tab      \tab     \tab Outcome 1 \tab \cr
}
}
}
