% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionBasedOnCI.R
\name{MakeDecisionBasedOnCI}
\alias{MakeDecisionBasedOnCI}
\title{MakeDecisionBasedOnCI}
\usage{
MakeDecisionBasedOnCI(dLowerCI, dUpperCI, lAnalysis)
}
\arguments{
\item{dLowerCI}{= lower limit of CI}

\item{dUpper}{= upper limit of CI}

\item{lAnaysis}{must have lAnalysis$dTV, and lAnalysis$dMAV}
}
\value{
Return a list of ( nSucces, nNoGo, nPause )
}
\description{
{This function  will make the Go/NoGo decision for an outcome given the CI and the MAV, TV }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionBasedOnCI.R}{View Code on GitHub} }
}
