% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionBasedOnPostProb.R
\name{MakeDecisionBasedOnPostProb.default}
\alias{MakeDecisionBasedOnPostProb.default}
\title{MakeDecisionBasedOnPostProb.default}
\usage{
\method{MakeDecisionBasedOnPostProb}{default}(cAnalysis, lCalcs)
}
\arguments{
\item{cAnalysis}{The analysis object.}

\item{lCalcs}{a list with dPrGrtTV, dPrGrtMAV, dMAVCutoff, dTVCutoff}
}
\value{
Return List of ( nGo, nNoGo, nPause ).  Only one of the elements should be 1 to reflect the decision.
}
\description{
{This function will make the Go/NoGo decision based on two posterior probabilities
Pr( \eqn{\theta} > MAV |data ) and Pr( \eqn{\theta} > TV | data ) where \eqn{\theta} is the
parameter of interest.
\tabular{cccc}{
          \tab Pr( \eqn{\theta} > MAV ) > dMAVCutoff \tab | \tab Pr( \eqn{\theta} > MAV) <= dMAVCutoff \cr
          \tab ------------------------------------- \tab | \tab -------------------------------------\cr
    Pr( \eqn{\theta} > TV ) > dTVCutoff   \tab Go    \tab | \tab  Pause \cr
          \tab ------------------------------------- \tab | \tab -------------------------------------\cr
    Pr( \eqn{\theta} > TV ) <= dTVCutoff  \tab Pause \tab | \tab No Go \cr
}
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionBasedOnPostProb.R}{View Code on GitHub} }
}
