% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionBasedOnPostProb.R
\name{MakeDecisionBasedOnPostProb.MAVOnly}
\alias{MakeDecisionBasedOnPostProb.MAVOnly}
\title{MakeDecisionBasedOnPostProb.MAVOnly}
\usage{
\method{MakeDecisionBasedOnPostProb}{MAVOnly}(cAnalysis, lCalcs)
}
\arguments{
\item{cAnalysis}{The analysis object.}

\item{lCalcs}{a list with  dPrGrtMAV, dPUpperCutoff, dLowerCutoff}
}
\value{
Return List of ( nGo, nNoGo, nPause ).  Only one of the elements should be 1 to reflect the decision.
}
\description{
{This function will make the Go/NoGo decision based on one posterior probability,
Pr( \eqn{\theta} > MAV |data ) where \eqn{\theta} is the
parameter of interest. Based on this probability the following decisions are made
\tabular{lc}{
\tab Decision \cr
Pr( \eqn{\theta} > MAV |data ) > dPUpperCutoff     \tab Go   \cr
Pr( \eqn{\theta} > MAV |data ) < dLowerCutoff      \tab No Go \cr
Otherwise                                          \tab Pause \cr
}
}
}
