% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionBasedOnPostProb.R
\name{MakeDecisionBasedOnPostProb}
\alias{MakeDecisionBasedOnPostProb}
\title{MakeDecisionBasedOnPostProbb}
\usage{
MakeDecisionBasedOnPostProb(cAnalysis, lCalcs)
}
\description{
{This function makes a Go/No Go decision based on posterior probabilities.   }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionBasedOnPostProb.R}{View Code on GitHub} }
}
