% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.AtLeastOne}
\alias{MakeDecisionDoses.AtLeastOne}
\title{MakeDecisionDoses.AtLeastOne}
\usage{
\method{MakeDecisionDoses}{AtLeastOne}(lDoseDec)
}
\description{
{This option allows for mulitiple doses and if all doses reach a No Go decision
then the combined decision is a No Go.  If ANY dose is a Go then a Go decision is made.     }
}
