% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.default}
\alias{MakeDecisionDoses.default}
\title{MakeDecisionDoses.default}
\usage{
\method{MakeDecisionDoses}{default}(lDoseDec)
}
\description{
{Generic function to make decisions when doses are present in an ISA.
No default defined so the simulations stop if a specific approach is not defined.  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionDoses.R}{View Code on GitHub} }
}
