% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.HighDoseOnly}
\alias{MakeDecisionDoses.HighDoseOnly}
\title{MakeDecisionDoses.HighDoseOnly}
\usage{
\method{MakeDecisionDoses}{HighDoseOnly}(lDoseDec)
}
\description{
{"This option allows for mulitiple doses but only uses the high dose for decision making "  }
}
