% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.MakeDecisionMCPMod}
\alias{MakeDecisionDoses.MakeDecisionMCPMod}
\title{MakeDecisionDoses.MakeDecisionMCPMod}
\usage{
\method{MakeDecisionDoses}{MakeDecisionMCPMod}(lDoseDec)
}
\description{
MakeDecisionDoses.MakeDecisionMCPMod
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionDoses.R}{View Code on GitHub} }
}
