% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses}
\alias{MakeDecisionDoses}
\title{MakeDecisionDoses}
\usage{
MakeDecisionDoses(lDoseDec)
}
\description{
{This function provides options for making decisons when doses are included in an ISA. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionDoses.R}{View Code on GitHub} }
}
