% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrialDecision.R
\name{MakeTrialDecision}
\alias{MakeTrialDecision}
\alias{MakeTrialDecision.default}
\title{MakeTrialDecision}
\usage{
\method{MakeTrialDecision}{default}(cISADesigns, lResAnalysis, vISAStatus,
  vIsFinalISAAnalysis, cRandomizer)
}
\description{
{This function determines the status of a trial (all ISAs) based on the input. Each ISA
can be in one of eight states. The following list provides the details of the returned vISAStatus and what
they mean.
\describe{
\item{Status 0}{The ISA is not open, the virtual trial is at a point prior to the ISA opening.}
\item{Status 1}{The ISA is open for enrollment.}
\item{Status 2}{The ISA has met the maximum enrollment.}
\item{Status 3}{The ISA was closed with a Go BEFORE Final Analysis (FA).  This can be a Final status for an ISA because it was closed.}
\item{Status 4}{The ISA was closed with a No Go BEFORE FA.  This can be a Final status for an ISA because it was closed.}
\item{Status 5}{The ISA was closed with a Go at FA.  This can be a Final status for an ISA because it was closed.}
\item{Status 6}{The ISA was closed with a No Go at FA.  This can be a Final status for an ISA because it was closed.}
\item{Status 7}{The ISA was closed with a Pause at FA.  This can be a Final status for an ISA because it was closed.}
}
The Status options of 3 or 4 indicate the ISA was closed early, before the FA, either for a Go or a No Go.  The options 5-7
all were closed at the FA.
}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeTrialDecision.R}{View Code on GitHub} }
}
