% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connection.R
\name{new_connection_form_postgres}
\alias{new_connection_form_postgres}
\title{New Connection Form - Postgres}
\usage{
new_connection_form_postgres()
}
\value{
A shiny tag list for inputing new connection details.
}
\description{
New Connection Form - Postgres
}
