% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{not_all_na}
\alias{not_all_na}
\title{Function to identify non-empty columns.}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical.
}
\description{
Function to identify non-empty columns.
}
\examples{
data.frame(x = c(1, 1), y = c(NA, NA)) \%>\% dplyr::select_if(not_all_na)
}
