% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{not_all_same}
\alias{not_all_same}
\title{Function to identify non-constant columns (NA is treated as a unique value).}
\usage{
not_all_same(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical.
}
\description{
Function to identify non-constant columns (NA is treated as a unique value).
}
\examples{
data.frame(x = c(1, 1), y = c(1, NA)) \%>\% dplyr::select_if(not_all_same)
}
