% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Returns TRUE if x is not in y.}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Anything.}

\item{y}{Anything.}
}
\value{
Logical.
}
\description{
Returns TRUE if x is not in y.
}
\examples{
"a" \%notin\% c("b", "c")
}
