% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{obj_to_str}
\alias{obj_to_str}
\title{Convert an object to a string.}
\usage{
obj_to_str(obj)
}
\arguments{
\item{obj}{An object.}
}
\value{
A string.
}
\description{
Convert an object to a string.
}
\examples{
data.frame(x = c(NA, "A string"), y = 1:2) \%>\% obj_to_str()
}
