% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCTOPUS-package.R
\docType{package}
\name{OCTOPUS-package}
\alias{OCTOPUS-package}
\title{OCTOPUS: OCTOPUS}
\description{
This package is used to simulate platform clinical trials. For more about platform trials a reference list is provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kwathen/OCTOPUS}
}

}
\author{
\strong{Maintainer}: J. Kyle Wathen \email{KyleWathen@gmail.com}

}
\keyword{internal}
