% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.nb.R
\name{octopus.glm.nb}
\alias{octopus.glm.nb}
\title{Fit a Negative Binomial Generalized Linear Model}
\usage{
octopus.glm.nb(formula, data, factors, specs4lsmeans = NULL)
}
\arguments{
\item{formula}{formula}

\item{data}{normalized reads, expecting gene as column, sample as row}

\item{factors}{factors, expecting factor name as column, sample as row}

\item{specs4lsmeans}{specs for lsmeans}
}
\value{
model if \code{formula} has respond y; flattened anova table, lsm table and se table if formula starts with ~.
}
\description{
Fit glm.nb , add anova to result, add lsm to result.\cr
Before fitting, \code{data} and \code{factors} will be merged by row.names, mismatching rows will be removed.\cr
If formula start with ~, will fit model with each data column of data as respond y.
}
\examples{
{
    data(sample_keys) # facotrs
    for(cur_col in colnames(sample_keys)){ sample_keys[,cur_col] <- as.factor(sample_keys[,cur_col]) }

    data(sample_reads) # reads
    # reads normalization
    norm.reads <- octopus.normalize(sample_reads)

    print(row.names(sample_keys))
    print(row.names(norm.reads))
    # transpose it to meet input requirements of founction octopus.glm.nb()
    norm.reads <- data.frame(t(norm.reads))
    print(row.names(norm.reads))

    # fit model with first gene column, column name is Bcin01g00040.1
    result <- octopus.glm.nb(Bcin01g00040.1 ~ Experiment + Experiment/GrowingFlat + Experiment/GrowingFlat/AgarFlat + Isolate + HostGenotype + HostGenotype*Isolate
                             ,norm.reads
                             ,sample_keys)
    print(result$anova)

    # calculate LSMean
    result$lsmeans <- lsmeans(result,~ Isolate | HostGenotype)
    summary(result$lsmeans)

    # fit model and pull LSMean in one call
    result <- octopus.glm.nb(Bcin01g00040.1 ~ Experiment + Experiment/GrowingFlat + Experiment/GrowingFlat/AgarFlat + Isolate + HostGenotype + HostGenotype*Isolate
                             ,norm.reads
                             ,sample_keys
                             ,specs4lsmeans = ~ Isolate | HostGenotype)
    print(result$anova)
    summary(result$lsmeans)

    # iterate through all data
    result <- octopus.glm.nb( ~ Experiment + Experiment/GrowingFlat + Experiment/GrowingFlat/AgarFlat + Isolate + HostGenotype + HostGenotype*Isolate
                              ,norm.reads
                              ,sample_keys
                              ,specs4lsmeans = ~ Isolate | HostGenotype)
    print(result$anova)
    print(result$lsm)
    print(result$se)

  }


}

