% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.nb.R
\name{octopus.normalize}
\alias{octopus.normalize}
\title{Normalize}
\usage{
octopus.normalize(x, dist_dir = "results/")
}
\arguments{
\item{x}{number matrix, expecting gene as row, sample as column}

\item{dist_dir}{}
}
\value{
number matrix
}
\description{
NA will be treated as 0
}
\examples{
tutorial("octopus.glm.nb") # to print tutorial
}
\seealso{
\link[octopus]{octopus.glm.nb}
}
