% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_reads.R
\name{octopus.short_reads}
\alias{octopus.short_reads}
\title{Extract Short Reads}
\usage{
octopus.short_reads(seq_files, references, ..., type = "single",
  dist_dir = "results/")
}
\arguments{
\item{seq_files}{Sequencing files , accepts .fastq or .gz format for files.\cr
If \code{seq_files} is a \code{List} or \code{Vector}, index of \code{seq_files} is assumed to be sampe names of sequencing files.\cr
If \code{seq_files} is a \code{data.frame}, row.names is assumed to be sampe names of sequencing files, the first column is assumed to be sequencing files, when \code{type} is \code{paired} the second column is assumed to be the second mate pair sequences.}

\item{references}{A comma-separated list of FASTA files containing the reference sequences to be aligned to}

\item{...}{Additional arguments to be passed on to the binaries. See ... of \link[Rbowtie]{bowtie}}

\item{type}{Could be one of c("single", "paired", "crossbow").\cr
If single, the input sequences are interpreted as single reads.\cr
If paired, they are supposed to be mate pair reads.\cr
If crossbow, they are considered to be Crossbow-style reads.}

\item{dist_dir}{folder for result file orig.reads.csv and orig.reads.info.csv}
}
\description{
Extract short reads from \code{seq_files} using \href{http://bowtie-bio.sourceforge.net/manual.shtml}{bowtie}.\cr
Output 2 files, \code{dist_dir}/orig.reads.csv and \code{dist_dir}/orig.reads.info.csv \cr
Remove \code{dist_dir}/reusing/ folder if don't want reuse data from that folder \cr
Setup ./tmp/ as an RAM disk folder will avoid lots of disk IOs, speed things up and protect you SSD.\cr
Example for linux :
\itemize{
  \item cd to curent folder.
  \item create ram disk in console : mount -t tmpfs -o size=4g tmpfs ./tmp/
}
Example for mac :
\itemize{
  \item cd to curent folder.
  \item mkdir -p tmp
  \item sudo mount -t tmpfs -o size=4096M tmpfs ./tmp/
}
For windows, there are a number of RAM disk softerwares you can use.
}
\examples{
{

    # single
    references <- "seq_data/cdna/Arabidopsis_thaliana.TAIR10.25.cdna.all.fa"
    seq_files <- data.frame(seq_file=c("seq_data/1_AACGTGAT_L003_R1_001.fastq.gz","seq_data/1_AACGTGAT_L007_R1_001.fastq","seq_data/3_AACGTGAT_L003_R1_001.fastq.gz")
                            ,sample_name=c("sample1","sample2","sample3")
                            ,stringsAsFactors = FALSE)
    row.names(seq_files) <- seq_files$sample_name

    octopus.short_reads(seq_files,references
                        ,p=3 # number of alignment threads to launch
                        ,`phred33-quals`=TRUE # input quals are Phred+33
                        ,t=TRUE # print wall-clock time taken by search phases
                        ,quiet=TRUE # print nothing but the alignments
                        ,trim5=10 # trim <int> bases from 5' (left) end of reads
    )

    # paired

    references <- "seq_data/cdna/Arabidopsis_thaliana.TAIR10.25.cdna.all.fa"
    seq_files <- data.frame(seq_file=c("seq_data/A9_S1_L001_R1_001.fastq.gz","seq_data/xxx_R1_001.fastq.gz","seq_data/xxxx_R1_001.fastq.gz")
                            ,seq_file_pair=c("seq_data/A9_S1_L001_R2_001.fastq.gz","seq_data/xxx_R2_001.fastq.gz","seq_data/xxxx_R2_001.fastq.gz")
                            ,sample_name=c("sample1","sample2","sample3")
                            ,stringsAsFactors = FALSE
                            )
    row.names(seq_files) <- seq_files$sample_name

    octopus.short_reads(seq_files,references
                        ,p=3 # number of alignment threads to launch
                        ,`phred33-quals`=TRUE # input quals are Phred+33
                        ,t=TRUE # print wall-clock time taken by search phases
                        ,quiet=TRUE # print nothing but the alignments
                        ,trim5=10 # trim <int> bases from 5' (left) end of reads
                        # ,y=TRUE # more sensitive but much slower, see http://bowtie-bio.sourceforge.net/manual.shtml#bowtie-options-y
                        ,type="paired"
                        ,dist_dir="results_paired/"
    )

    # multiple referencing files
    references  <- c("seq_data/ref_sequences/Botrytisfusarivirus1.txt","seq_data/ref_sequences/BotrytisHypovirus1.txt")
    seq_files <- data.frame(seq_file=c("seq_data/1_AACGTGAT_L003_R1_001.fastq.gz","seq_data/1_AACGTGAT_L007_R1_001.fastq","seq_data/3_AACGTGAT_L003_R1_001.fastq.gz")
                            ,sample_name=c("sample1","sample2","sample3")
                            ,stringsAsFactors = FALSE)
    row.names(seq_files) <- seq_files$sample_name

    octopus.short_reads(seq_files,references
                        ,p=3 # number of alignment threads to launch
                        ,`phred33-quals`=TRUE # input quals are Phred+33
                        ,t=TRUE # print wall-clock time taken by search phases
                        ,quiet=TRUE # print nothing but the alignments
                        ,trim5=10 # trim <int> bases from 5' (left) end of reads
    )


  }


}

