% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlatformTrialSimulator.R
\docType{package}
\name{PlatformTrialSimulator}
\alias{PlatformTrialSimulator}
\alias{PlatformTrialSimulator-package}
\title{Platform Trial Simulator}
\description{
This package is used to simulate platform clinical trials.  Platform trials are clinical trials that are designed with the flexibility to add
new interventions as the trial progresses.
}
\section{License}{

This package is governed by the JRD Platform Trial Simulation License, which is the GNU General Public License V3 with additional terms. The precise license terms are located in the files LICENSE and GPL.

To view license you may use the following command:

file.show( system.file("LICENSE", package="PlatformTrialSimulatorP"))

file.show( system.file("GPL", package="PlatformTrialSimulatorP"))
}

\section{Introduction}{

General intro to the packages ..
}

\section{Basic description}{

Stuff
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kwathen/PlatformTrialSimulator}
}

}
\author{
\strong{Maintainer}: J. Kyle wathen \email{kylewathen@gmail.com}

}
