\name{PlatformTrialSimulator-package}
\alias{PlatformTrialSimulator-package}
\docType{package}
\title{Platform Trial Simulator}
\description{Full description to go here}
\details{
\tabular{ll}{
Package: \tab PlatformTrialSimulator\cr
Type: \tab Package\cr
Version: \tab 1.2.1.7\cr
Date: \tab 01-01-2019\cr
License: \tab See files LICENSE and GPL\cr
\cr
\tab This package is governed by the JRD Platform Trial Simulation License, which is the GNU General Public License V3 with additional terms. The precise license terms are located in the files LICENSE and GPL.\cr
\cr
\tab To view license you may use the following command:\cr
\tab file.show( system.file("LICENSE", package="PlatformTrialSimulatorP"))\cr
\tab file.show( system.file("GPL", package="PlatformTrialSimulatorP"))\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{J. Kyle Wathen}



