% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentComparitor.R
\name{PlotPlatformRecruitment}
\alias{PlotPlatformRecruitment}
\title{PlotPlatformRecruitment}
\usage{
PlotPlatformRecruitment(nQtyReps, nMaxQtyPats, vPatsPerMonthPerSite1,
  dDelayToStartPOC, dDelayBetweenTrialsPOC, vQtySitesPOC,
  dDelayToStartPlat, dDelayBetweenTrialsPlat, vQtyOfSitesPlat1,
  vQtyOfSitesPlat2)
}
\description{
{This function creates a plot of recruitment comparing a platform with 2 ISAs and 2 consecutive POC studies.
Mainly used by the Shiny app execute by calling RunExample( "CompareRecruitment" ). }
}
