% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{ProcessData.ProcessReptMeas}
\alias{ProcessData.ProcessReptMeas}
\title{ProcessData.ProcessReptMeas}
\usage{
\method{ProcessData}{ProcessReptMeas}(cDS, dCurrentTime)
}
\description{
ProcessData.ProcessReptMeas
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/ProcessData.R}{View Code on GitHub} }
}
