% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{ProcessData.ProcessSingleTimeOutcome}
\alias{ProcessData.ProcessSingleTimeOutcome}
\title{ProcessData.ProcessSingleTimeOutcome}
\usage{
\method{ProcessData}{ProcessSingleTimeOutcome}(cDS, dCurrentTime)
}
\description{
ProcessData.ProcessSingleTimeOutcome
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/ProcessData.R}{View Code on GitHub} }
}
