% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{ProcessData}
\alias{ProcessData}
\title{ProcessData}
\usage{
ProcessData(cDS, dCurrentTime)
}
\description{
{ This function is used to process the data before the RunAnalysis function is called.  This can
be useful for cases where the data is simulated/collected in one for but another form is needed for analysis.  For
example, the patient data over time is simulated but the data is analysed as change from baseline. }
}
