% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizeBetweenISA.R
\name{RandomizeBetweenISA.default}
\alias{RandomizeBetweenISA.default}
\title{RandomizeBetweenISA.default}
\usage{
\method{RandomizeBetweenISA}{default}(cRandomizer, vISAStatus,
  dCurrentTime)
}
\description{
{RandomizeBetweenISA.default is not implemented to avoid potential use when the user
misspells the desired ISA randomizer.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeBetweenISA.R}{View Code on GitHub} }
}
