% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizeBetweenISA.R
\name{RandomizeBetweenISA.EqualRandomizer}
\alias{RandomizeBetweenISA.EqualRandomizer}
\title{RandomizeBetweenISA.EqualRandomizer}
\usage{
\method{RandomizeBetweenISA}{EqualRandomizer}(cRandomizer, vISAStatus, dCurrentTime)
}
\description{
{For all ISAs with vISAStus = 1, select an ISA index with equal probability.  }
}
