% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizeBetweenISA.R
\name{RandomizeBetweenISA}
\alias{RandomizeBetweenISA}
\title{RandomizeBetweenISA}
\usage{
RandomizeBetweenISA(cRandomizer, vISAStatus, dCurrentTime)
}
\description{
{ Generic method to randomize between ISAs.   Any version of
RandomizeBetweenISA should just return the index of the ISA the patient is assigned to.  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeBetweenISA.R}{View Code on GitHub} }
}
