% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizeWithinISA.R
\name{RandomizeWithinISA}
\alias{RandomizeWithinISA}
\alias{RandomizeWithinISA.default}
\alias{RandomizeWithinISA.EqualRandomizer}
\alias{RandomizeWithinISA.POCRandomizer}
\alias{RandomizeWithinISA.DelayedStartRandomizer}
\title{RandomizeWithinISA}
\usage{
RandomizeWithinISA(cISARand, dCurrentTime)

\method{RandomizeWithinISA}{default}(cISARand, dCurrentTime)

\method{RandomizeWithinISA}{EqualRandomizer}(cISARand, dCurrentTime)

\method{RandomizeWithinISA}{POCRandomizer}(cISARand, dCurrentTime)

\method{RandomizeWithinISA}{DelayedStartRandomizer}(cISARand, dCurrentTime)
}
\description{
{ This function is used in close connection with InitializeISARandomizer and
if a new version of this fuction is developed a new version InitializeISARandomizer with the same
class is required.  RandomizeWithInISA should return a list with at least the following items
 nTrt - The treatment within the ISA that the patient should receive
 cISARand - the same is cISARand but if it is updates this allows the caller to get the update. }
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: { Because several options are provided and there in no well defined default
an stop error occurs if you call the default method. }

\item \code{EqualRandomizer}: { Patients are randomized equally/fairly amount the treatments in
the ISA. }

\item \code{POCRandomizer}: { This randomzier will randomize the first patients to a select
set of arms or doses in the ISA (eg a POC phase).  The remaining arms or doses are opened based on the number of intitial patients. }

\item \code{DelayedStartRandomizer}: { This version will randomize the first patients to a select set of arms or doses in the ISA (eg a POC phase)
The remaining arms or doses are opened based on the provided times to open. }
}}

\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeWithinISA.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeWithinISA.R}{View Code on GitHub} }

InitializeISARandomizer.EqualRandomizer

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeWithinISA.R}{View Code on GitHub} }

InitializeISARandomizer.POCRandomizer

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeWithinISA.R}{View Code on GitHub} }

InitializeISARandomizer.DelayedStartRandomizer

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RandomizeWithinISA.R}{View Code on GitHub} }
}
