% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_db_credentials.R
\name{retrieve_db_credentials}
\alias{retrieve_db_credentials}
\title{Retrieve DB credentials}
\usage{
retrieve_db_credentials()
}
\value{
A character vector of connection names.
}
\description{
This function saves DB credentials on the local computer.
}
