% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Octopus App}
\usage{
run_app(con, options = list())
}
\arguments{
\item{con}{A database connection object. The result of DBI::dbConnect().}

\item{options}{A named list of options to be passed along to shinyApp().}
}
\value{
An RShiny instance.
}
\description{
This function runs the app.
}
