% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.AOV.R
\name{RunAnalysis.AOV}
\alias{RunAnalysis.AOV}
\title{RunAnalysis.AOV}
\usage{
\method{RunAnalysis}{AOV}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis, cRandomizer)
}
\description{
{This is the method used for running the Analysis of Variance (AOV) .
Specifically, lm( vOut ~ vBaseline + vTrt,na.action = na.omit)"  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunAnalysis.AOV.R}{View Code on GitHub} }
}
