% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.BayesianNormal.R
\name{RunAnalysis.BayesianNormal}
\alias{RunAnalysis.BayesianNormal}
\title{RunAnalysis.BayesianNormal}
\usage{
\method{RunAnalysis}{BayesianNormal}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\description{
{This is the method used for running the analysis of a BayesianNormal normal model.}
}
