% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.R
\name{RunAnalysis.default}
\alias{RunAnalysis.default}
\title{RunAnalysis.default}
\usage{
\method{RunAnalysis}{default}(
  cAnalysis,
  lDataAna,
  nISAAnalysisIndx,
  bIsFinalISAAnalysis,
  cRandomizer
)
}
\description{
{This is the method used for running the analysis as a default.  This method is not defined and is
in place in-case someone provides an invalid class type for RunAnalysis.  This version will stop execution.  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunAnalysis.R}{View Code on GitHub} }
}
