% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.Ranked.R
\name{RunAnalysis.Ranked}
\alias{RunAnalysis.Ranked}
\title{RunAnalysis.Ranked}
\usage{
\method{RunAnalysis}{Ranked}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\description{
{This is the method used for running the Ranked analysis. Specifically, wilcox.test with exact = TRUE"  }
}
