% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.R
\name{RunAnalysis}
\alias{RunAnalysis}
\title{RunAnalysis}
\usage{
RunAnalysis(cAnalysis, lDataAna, nISAAnalysisIndx, bIsFinalISAAnalysis,
  cRandomizer)
}
\arguments{
\item{cAnalysis}{The class( cAnalysis ) determines the specific version of RunAnalysis that is called. It contains
the details about the analysis such as the priors, MAV, TV, decsiion cutoff boundaries.}

\item{lDataAna}{The data that is used int he analysis.  Typically contains vISA (the ISA for the patient),
vTrt (treatment for each patient), vOut (the outcome for each patient)}

\item{nISAAnalysisIndx}{index of the analysis used for changing boundaries)}

\item{cRandomizer}{The randomizer, mainly used for cases with covariates. For most cases without covariates,
simply add cRandomizer = cRandomizer to the return list.}

\item{bIsFinaISAAnalysis}{TRUE or FALSE, often we change the value of the cutoff at the final analysis for an ISA}
}
\value{
Each Analysis must return a list with at least the following list( nGo = nGo, nNoGo= nNoGo,  nPause = nPause, cRandomizer = cRandomizer ).
The list is a list of indicator that only one of which = 1, the others are 0
For example, if you did a Go/NoGo with a CI see the results of MakeDecisionBasedOnCI()
If the analysis return more info that is fine.
}
\description{
{This is the method used for running the analysis.  There are several options available for different types
of data.
Inputs:
    cAnalysis - The class( cAnalysis ) determines the specific version of RunAnalysis that is called. It contains
                the details about the analysis such as the priors, MAV, TV, decsiion cutoff boundaries.
    lDataAna  - The data that is used int he analysis.  Typically contains vISA (the ISA for the patient),
                vTrt (treatment for each patient), vOut (the outcome for each patient)

Return:   Each Analysis must return a list with at least the following
 list( nGo = nGo, nNoGo= nNoGo,  nPause = nPause)
  The list is a list of indicator that only one of which = 1, the others are 0
For example, if you did a Go/NoGo with a CI see the results of MakeDecisionBasedOnCI()
If the analysis return more info that is fine.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunAnalysis.R}{View Code on GitHub} }
}
