% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.ReptMeasAnalysis.R
\name{RunAnalysis.ReptMeasAnalysis}
\alias{RunAnalysis.ReptMeasAnalysis}
\title{RunAnalysis.ReptMeasAnalysis}
\usage{
\method{RunAnalysis}{ReptMeasAnalysis}(cAnalysis, lDataAna,
  nISAAnalysisIndx, bIsFinalISAAnalysis, cRandomizer)
}
\description{
{This is the method used for running the Repeated Measure Analysis .
Specifically, gls(vOut ~ vBaseline + vTrt + vTime + vTrt:vTime,
weights = varIdent(form = ~1|vTime),
correlation = corSymm(form=~1|vIND), na.action = na.omit )"  }
}
