% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.TTest.R
\name{RunAnalysis.TTest}
\alias{RunAnalysis.TTest}
\title{RunAnalysis.TTest}
\usage{
\method{RunAnalysis}{TTest}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis, cRandomizer)
}
\description{
{This is the method used for running the Repeated Measure Analysis .
Specifically,
t.test(lDataAna$vOut ~ as.factor(lDataAna$vTrt), alternative = c("two.sided"),mu = 0, exact = NULL, correct = TRUE,  conf.int = TRUE, conf.level = dCILevel)"  }
}
