% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{RunExample}
\alias{RunExample}
\title{RunExample}
\usage{
RunExample(example)
}
\description{
{This function is used to run the example in the package.  For a current list of example call
  PlatformTrialSimulator::RunExample() and you will get a list of the available examples.

  There are example(s) that require shiny so you must install the shiny package to run some of the examples.
  }
}
\examples{
  PlatformTrialSimulator::RunExample( "CompareRecruitment" )
}
