% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFinalTrialAnalysis.R
\name{RunFinalAnalysis}
\alias{RunFinalAnalysis}
\alias{RunFinalTrialAnalysis}
\title{RunFinalAnalysis}
\usage{
RunFinalTrialAnalysis(cISADesigns, cEnrolledPats, vISAStatus, vISAAnalysisIndx)
}
\description{
{ Need description }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunFinalTrialAnalysis.R}{View Code on GitHub} }
}
