% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunISAAnalysis.R
\name{RunISAAnalysis.default}
\alias{RunISAAnalysis.default}
\title{RunISAAnalysis.default}
\usage{
\method{RunISAAnalysis}{default}(cISAAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\description{
The cISAAnalysis object has a list of Analysis
}
