% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunISAAnalysis.R
\name{RunISAAnalysis}
\alias{RunISAAnalysis}
\title{RunISAAnalysis}
\usage{
RunISAAnalysis(cISAAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis, cRandomizer)
}
\description{
{Generic method called to run the ISA analysis.
Because the decision is tied to the analysis, each RunISAAnalysis should contain the elements nGo, nNoGo, nPause as part ot the list retured.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunISAAnalysis.R}{View Code on GitHub} }
}
