% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRealTrialAnalysis.R
\name{RunRealTrialAnalysis.default}
\alias{RunRealTrialAnalysis.default}
\title{RunRealTrialAnalysis.default}
\usage{
\method{RunRealTrialAnalysis}{default}(cTrialDesign, strInputFile,
  strOutputFile, nSeed)
}
\description{
{RunRealTrialAnalysis.default provides the default implmentation as it is not anticipated to be utilized very often.  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunRealTrialAnalysis.R}{View Code on GitHub} }
}
