% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTrialAnalysis.R
\name{RunTrialAnalysis}
\alias{RunTrialAnalysis}
\alias{RunTrialAnalysis.default}
\title{RunTrialAnalysis}
\usage{
RunTrialAnalysis(
  cISADesigns,
  cEnrolledPats,
  vISAStatus,
  dCurrentTime,
  vRunISAAnalysis,
  vISAAnalysisIndx,
  vIsFinalISAAnalysis,
  cRandomizer
)

\method{RunTrialAnalysis}{default}(
  cISADesigns,
  cEnrolledPats,
  vISAStatus,
  dCurrentTime,
  vRunISAAnalysis,
  vISAAnalysisIndx,
  vIsFinalISAAnalysis,
  cRandomizer
)
}
\description{
{ RunTrialAnalysis is a generic method to execute the trial level analysis, eg execute all ISA analysis.
Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA   }

{ RunTrialAnalysis is a generic method to execute the trial analysis.
Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA.
Trial level analysis go through the following steps:
    1. CreateDataSetForAnalysis - creates the data set for analysis for all ISAs will go through and call ProcessData
       based on  the specific class of the outcome
    2. Loop through each ISA and build the dataset for analysis for the specific ISA if it is borrowing.  The
       borrowing is done via the SubsetData call and it will depend on the specific class of cISADesigns[[i]]$cISAAnalysis. The
       available options are AllControls or NoBorrowing but since SubsetData is an S3 generic new functions can be added.
    3. Runs the analysis for each ISA via a call to  RunISAAnalysis

  Because the decision is tied to the analysis, each ISA analysis should contain the elements nGo, nNoGo, nPause as part of the list returned.
 }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunTrialAnalysis.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunTrialAnalysis.R}{View Code on GitHub} }
}
