% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SelectList}
\alias{SelectList}
\title{SelectList}
\usage{
SelectList(lData, lValue)
}
\arguments{
\item{lData}{- The dataset (as a list) that you want to select from}

\item{lValue}{- A list with the desired values for each column in lData}
}
\description{
{Determine if the elements of lData have the values of lValue, this is like SQL select statement with AND for all values}
}
