% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridFunctions.R
\name{SetRunningEnvironment}
\alias{SetRunningEnvironment}
\title{SetRunningEnvironment}
\usage{
SetRunningEnvironment(job.id, cmdArgs)
}
\description{
{This function is used for setting the running environment and output file names accordingly. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/GridFunctions.R}{View Code on GitHub} }
}
