% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPatientOutcomes.R
\name{SimPatientOutcomes.Correlated}
\alias{SimPatientOutcomes.Correlated}
\title{SimPatientOutcomes}
\usage{
\method{SimPatientOutcomes}{Correlated}(cSimOutcomes, cISADesign)
}
\description{
The cSimOutcomes object has a list of sim outcomes and this version will utilize the
first outcome to simulate all outcomes correlated.  Each simoutcome must have an attribute vColIndex that
specifies which columns out of the matrix will be used for that outcome.
}
