% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPatientOutcomes.R
\name{SimPatientOutcomes.Independent}
\alias{SimPatientOutcomes.Independent}
\title{SimPatientOutcomes}
\usage{
\method{SimPatientOutcomes}{Independent}(cSimOutcomes, cISADesign)
}
\description{
The cSimOutcomes object has a list of sim outcomes and this version will go through and call the
SimPatientOutcomes for each type.  Each outcome is simulated independently.
}
