% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPatientOutcomes.MVN.R
\name{SimPatientOutcomes.MVN}
\alias{SimPatientOutcomes.MVN}
\title{SimPatientOutcomes.MVN}
\usage{
\method{SimPatientOutcomes}{MVN}(cSimOutcomes, cISADesign, lPatCovISA)
}
\description{
Simulate a MVN outcomes.  This one will likely be used by outcomes that are simulated as correlated.
}
