% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPatientOutcomes.Correlated.R,
%   R/SimPatientOutcomes.Independent.R, R/SimPatientOutcomes.R
\name{SimPatientOutcomes.Correlated}
\alias{SimPatientOutcomes.Correlated}
\alias{SimPatientOutcomes.Independent}
\alias{SimPatientOutcomes}
\alias{SimPatientOutcomes.default}
\title{SimPatientOutcomes.Correlated}
\usage{
\method{SimPatientOutcomes}{Correlated}(cSimOutcomes, cISADesign, dfPatCovISA)

\method{SimPatientOutcomes}{Independent}(cSimOutcomes, cISADesign, dfPatCovISA)

SimPatientOutcomes(cSimOutcomes, cISADesign, dfPatCovISA)

\method{SimPatientOutcomes}{default}(cSimOutcomes, cISADesign, dfPatCovISA)
}
\description{
SimPatientOutcomes {This function is intended to simulate the outcomes for a given ISA.}
}
\section{Methods (by class)}{
\itemize{
\item \code{Correlated}: { The cSimOutcomes object has a list of sim outcomes and this version will utilize the
first outcome to simulate all outcomes correlated.  Each simoutcome must have an attribute vColIndex that
specifies which columns out of the matrix will be used for that outcome. }

\item \code{Independent}: { The cSimOutcomes object has a list of sim outcomes and this version will go through and call the
SimPatientOutcomes for each type.  Each outcome is simulated independently. }

\item \code{default}: { This function is intended to simulate the outcomes for a given ISA.
Because several options are provided and there in no well defined default
an stop error occurs if you call the default method. }
}}

\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimPatientOutcomes.Correlated.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimPatientOutcomes.Independent.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimPatientOutcomes.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimPatientOutcomes.R}{View Code on GitHub} }
}
