% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentComparitor.R
\name{SimulateAccrual}
\alias{SimulateAccrual}
\title{SimulateAccrual}
\usage{
SimulateAccrual(nQtyReps, nMaxQtyPats, dDelayBetweenTrials,
  vPatsPerMonthPerSite1, vQtyOfSites1, vPatsPerMonthPerSite2, vQtyOfSites2,
  dDelayToStart)
}
\description{
{This function simulates the accural for 2 ISAs.  Mainly used by the Shiny app execute by calling RunExample( "CompareRecruitment" ). }
}
