% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateAllPatientCovariates.R
\name{SimulateAllPatientCovariates}
\alias{SimulateAllPatientCovariates}
\alias{SimulateAllPatientCovariates.default}
\title{SimulateAllPatientCovariates}
\usage{
SimulateAllPatientCovariates(cSimCovariates, cTrialDesign)

\method{SimulateAllPatientCovariates}{default}(cSimCovariates,
  cTrialDesign)
}
\description{
SimulateAllPatientCovariates {This function is used to simulate all patient covariates in the platform.  }

SimulateAllPatientCovariates {This function is used to simulate all patient covariates in the platform.
This function is generic in case there is a need to override but
is implemented such that it should be suitable for most cases.  Specifically,
this function calls SimulatePatientCovariates for each ISA. }
}
