% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateAllPatientOutcomes.R
\name{SimulateAllPatientOutcomes}
\alias{SimulateAllPatientOutcomes}
\alias{SimulateAllPatientOutcomes.default}
\title{SimulateAllPatientOutcomes}
\usage{
SimulateAllPatientOutcomes(cScen, cTrialDesign, dfPatCov)

\method{SimulateAllPatientOutcomes}{default}(cScen, cTrialDesign, dfPatCov)
}
\description{
SimulateAllPatientOutcomes {This function is used to simulate all patient outcomes in the platform.}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: { This function is used to simulate all patient outcomes in the platform.
This function is generic in case there is a need to override but
is implemented such that it should be suitable for most cases.  Specifically,
this function calls SimPatientOutcomes for each ISA. }
}}

