% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccrualMethods.R
\name{SimulateArrivalTimes}
\alias{SimulateArrivalTimes}
\alias{SimulateArrivalTimes,AccrualMethods-method}
\title{SimulateArrivalTimes}
\usage{
\S4method{SimulateArrivalTimes}{AccrualMethods}(cAP)
}
\arguments{
\item{cAP}{An AccrualMethods object created by \code{\link{NewAccrualProcess}}}

\item{vQtyPatsPerMonth}{A vector (possibly of length 1 for constant rate) for the number of patient accrued each month}

\item{nMaxQtyPatients}{The maximum of of patient recruitment times to simulate.}

\item{nMaxMonthsOfAccrual}{The maximum number of months that you want to simulate accrual for.}
}
\description{
{This class can be used to simulate the arrival times of patients in a simulated clinical trial.  The function
NewAccrualProcess returns an object used to simulate arrival times for patients by calling SimulateArrivalTime.
The class assumes
 After calling NewAccrualProces you can call SimulateArrivalTimes to generate a vector of accrual times.
}
}
\details{
{ If both nMaxQtyPatients and nMaxMonthsOfAccrual are supplied then the returned vector will be at most
nMaxQtyPatients in length.  However, when nMaxMonthsOfAccrual all values will be less than nMaxMonthsOfAccrual so the
length of the returned vector could be less than nMaxQtyPatient.  }
}
\examples{
  vPatsPerMonth  <- c(5, 10, 15, 20, 30, 40, 50) # The ramp up in expected Pat/month
  nMaxQtyPats    <- 100  # Maximum of 100 patients
  ap             <- NewAccrualProcess( vQtyPatsPerMonth = vPatsPerMonth, nMaxQtyPatients = nMaxQtyPats )
  vAccTimes      <- SimulateArrivalTimes( ap )

}
\seealso{
\code{\link{NewAccrualProcess}} for creating the object.
}
