% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateISAStartTime.R
\name{SimulateISAStartTime.default}
\alias{SimulateISAStartTime.default}
\title{Simulate ISA Start Time}
\usage{
\method{SimulateISAStartTime}{default}(cParams)
}
\description{
This version should not be implemented
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimulateISAStartTime.R}{View Code on GitHub} }
}
