% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateISAStartTime.R
\name{SimulateISAStartTime.SetTime}
\alias{SimulateISAStartTime.SetTime}
\title{SimulateISAStartTime SetTime}
\usage{
\method{SimulateISAStartTime}{SetTime}(cParams)
}
\description{
This version assumes the cParams has a dParam1 which specifies the start time for the ISA, eg it is fixed when the ISA starts
}
