% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateISAStartTime.R
\name{SimulateISAStartTime.Uniform}
\alias{SimulateISAStartTime.Uniform}
\title{Simulate ISA Start Time from a uniform distribution}
\usage{
\method{SimulateISAStartTime}{Uniform}(cParams)
}
\description{
This version assumes the cParams has a dParam1 and dParam2 and simulates a unif( dParam1, dParam2 ) as the ISA start time
}
