% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateScenario.R
\name{SimulateScenario}
\alias{SimulateScenario}
\alias{SimulateScenario.default}
\title{SimulateScenario}
\usage{
SimulateScenario(cScen, cTrialDesign)

\method{SimulateScenario}{default}(cScen, cTrialDesign)
}
\description{
{ SimulateScenario This function is the main funciton in simulating a scenario.
The scenario will be simulated cScen$nQtyReps times.  }
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method that is suitable for most cases.
}}

\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimulateScenario.R}{View Code on GitHub} }

{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimulateScenario.R}{View Code on GitHub} }
}
