% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateSingleTrial.R
\name{SimulateSingleTrial}
\alias{SimulateSingleTrial}
\alias{SimulateSingleTrial.default}
\title{SimulateSingleTrial}
\usage{
\method{SimulateSingleTrial}{default}(cScen, cTrialDesign)
}
\description{
{This function simulates 1 virtual trial based on cScen and cTrialDesign.  This is the main function for running
the simulation.}
}
