% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{slice}
\alias{slice}
\title{Slice list into sublists by n}
\usage{
slice(x, n)
}
\arguments{
\item{x}{list}

\item{n}{desired length of sublist}
}
\value{
A list of sublists
}
\description{
Slice list into sublists by n
}
\examples{
{
    x <- 5:505
    slice(x,100) # 5 sublists
    slice(x,99)  # 6 sublists
    slice(x,101) # 4 sublists
  }



tutorial("slice") # print tutorial to console

}

