% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{str_to_obj}
\alias{str_to_obj}
\title{Convert a string (generated using obj_to_str) back to an object.}
\usage{
str_to_obj(str)
}
\arguments{
\item{str}{A string generated using obj_to_str().}
}
\value{
An object.
}
\description{
Convert a string (generated using obj_to_str) back to an object.
}
\examples{
data.frame(x = c(NA, "A string"), y = 1:2) \%>\% obj_to_str() \%>\% str_to_obj()
}
