% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData.AllControls}
\alias{SubsetData.AllControls}
\title{SubsetData.AllControls}
\usage{
\method{SubsetData}{AllControls}(cSimDataAna, nISA)
}
\description{
{ Include all control data}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SubsetData.R}{View Code on GitHub} }
}
