% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData}
\alias{SubsetData}
\title{SubsetData}
\usage{
SubsetData(cSimDataAna, nISA)
}
\description{
{This function is used to subset data.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SubsetData.R}{View Code on GitHub} }
}
