% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summaree}
\alias{summaree}
\title{Modificaton of summary() that reports strings as factors.}
\usage{
summaree(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A summary.
}
\description{
Modificaton of summary() that reports strings as factors.
}
\examples{
summaree(data.frame(x = LETTERS, y = 1:26))
}
