% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{tutorial}
\alias{tutorial}
\title{Print Tutorial}
\usage{
tutorial(topic)
}
\arguments{
\item{topic}{tutorial topic, it is usually a function name.}
}
\value{
tutorial code as a function
}
\description{
Print tutorial for function/class and return tutorial code as a function. It is a generic function.
}
\examples{
{
    # access tutorial code for class info
    tutorial("info")          # use class name string
    tutorial(create.info())   # use object

    tutorial(create.info)     # use function name
    tutorial("create.info")   # use function name string

    tutorial_func <- tutorial("info")
    tutorial_func() # run tutorial code, suggest read it first
  }



tutorial("tutorial") # print tutorial to console

}

