% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateRandomizer.R
\name{UpdateRandomizer.Outcome1Only}
\alias{UpdateRandomizer.Outcome1Only}
\alias{UpdateRandomizer.Outcome2Only}
\title{UpdateRandomizer.Outcome1Only}
\usage{
\method{UpdateRandomizer}{Outcome1Only}(lDecision, lResAnalysis,
  cRandomzier)

\method{UpdateRandomizer}{Outcome2Only}(lDecision, lResAnalysis,
  cRandomzier)
}
\description{
UpdateRandomizer.Outcome1Only { }

UpdateRandomizer.Outcome1Only { }
}
