% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_database.R
\name{view_database}
\alias{view_database}
\title{View Database Connection with Octopus}
\usage{
view_database(con, options = list(), max_file_upload_size = 2000 * 1024^2)
}
\arguments{
\item{con}{A database connection object. The result of DBI::dbConnect().}

\item{options}{A named list of options to be passed along to shinyApp().}

\item{max_file_upload_size}{An integer. The max number of bits allowed in file uploads.}
}
\value{
An R Shiny instance.
}
\description{
This function opens a shiny instance where the database can
  be viewed.
}
