% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{write_data}
\alias{write_data}
\title{Function to write dataframe to a delimited file (but only if it has changed).}
\usage{
write_data(data, path, delim = "\\t", silent = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{path}{A file path.}

\item{delim}{A delimiter. Default is tab delimited.}

\item{silent}{Logical.}
}
\value{
Nothing.
}
\description{
Function to write dataframe to a delimited file (but only if it has changed).
}
\examples{
library(magrittr)
data.frame(x = c(NA, "A string"), y = 1:2) \%>\% write_data("my_data.tsv")
}
