% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connection.R
\name{write_new_connection_postgres}
\alias{write_new_connection_postgres}
\title{Write New Connection - Postgres}
\usage{
write_new_connection_postgres(
  new_name,
  new_host,
  new_user,
  new_password,
  new_port,
  new_dbname
)
}
\arguments{
\item{new_name}{A string connection name.}

\item{new_host}{A string connection host. Ex: localhost}

\item{new_user}{A string connection username.}

\item{new_password}{A string connection password.}

\item{new_port}{An integer connection port.}

\item{new_dbname}{A string connection database name.}
}
\value{
A connections dataframe is returned.
The connection details are also written to a connection file.
}
\description{
Write New Connection - Postgres
}
