% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_connection.R
\name{write_table_mysql}
\alias{write_table_mysql}
\title{A Database Specific Function for Uploading Data Frames}
\usage{
write_table_mysql(con, schema, table_name, data, temporary = FALSE)
}
\arguments{
\item{con}{A database connection object.}

\item{schema}{A string containing the schema name.}

\item{table_name}{A string for the new table name.}

\item{data}{A data frame to be uploaded.}

\item{temporary}{A logical value. Should the table be temporary?}
}
\value{
A result string. Either "Success" or an error message.
}
\description{
A Database Specific Function for Uploading Data Frames
}
