% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddControlData.ProcessReptMeasChngBaseline.R
\name{AddControlData.ProcessReptMeasChngBaseline}
\alias{AddControlData.ProcessReptMeasChngBaseline}
\title{AddControlData.ProcessReptMeasChngBaseline}
\usage{
\method{AddControlData}{ProcessReptMeasChngBaseline}(cISAData, cSourceData,
  nISA)
}
\description{
{ Control data when processed as a Repeated measure change from outcome.}
}
