% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddControlData.ProcessSingleTimeOutcome.R
\name{AddControlData.ProcessSingleTimeOutcome}
\alias{AddControlData.ProcessSingleTimeOutcome}
\title{AddControlData.ProcessSingleTimeOutcome}
\usage{
\method{AddControlData}{ProcessSingleTimeOutcome}(cISAData, cSourceData, nISA)
}
\description{
{ Control data when processed as a single time point.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/AddControlData.ProcessSingleTimeOutcome.R}{View Code on GitHub} }
}
