% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CheckGlobalVariables}
\alias{CheckGlobalVariables}
\title{CheckGlobalVariables}
\usage{
CheckGlobalVariables()
}
\arguments{
\item{lData}{- The dataset (as a list) that you want to select from}

\item{lValue}{- A list with the desired values for each column in lData}
}
\description{
{Determine if the global variables that are needed have been defined and if not define them. Use of global variables
should be restricted to the variables that are used throughout the code and only in VERY limited use.}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/Functions.R}{View Code on GitHub} }
}
