% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckISAEnrollmentStatus.R
\name{CheckISAEnrollmentStatus.default}
\alias{CheckISAEnrollmentStatus.default}
\title{CheckISAEnrollmentStatus.default}
\usage{
\method{CheckISAEnrollmentStatus}{default}(cRandomizer, vISAStatus, dfCov)
}
\description{
CheckISAEnrollmentStatus.default { This function will check the cRandomzier to make sure if an ISA is open for the covariate group dfCov
Two cases
 If No covaraite - dfCov = NULL then  vEnrollmentStatus = vISAStatus
 If Covariates - dfCov != NULL then vEnrollmentStatus = vISAStatus but if an ISA does not enroll for dfCov the value will be 0, indicating it is not open.
 Essentially, if an ISA is open for dfCov the value in vRetISAStatus = 1, if not it will be a value != 1}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CheckISAEnrollmentStatus.R}{View Code on GitHub} }
}
