% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckISAEnrollmentStatus.R
\name{CheckISAEnrollmentStatus.default}
\alias{CheckISAEnrollmentStatus.default}
\title{CheckISAEnrollmentStatus.default}
\usage{
\method{CheckISAEnrollmentStatus}{default}(cRandomizer, vISAStatus, dfCov)
}
\description{
CheckISAEnrollmentStatus.default { This function will check the cRandomzier to make sure if an ISA is open for the covariate group dfCov}
}
