% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckISAEnrollmentStatus.R
\name{CheckISAEnrollmentStatus}
\alias{CheckISAEnrollmentStatus}
\title{CheckISAEnrollmentStatus}
\usage{
CheckISAEnrollmentStatus(cRandomizer, vISAStatus, dfCov)
}
\description{
CheckISAEnrollmentStatus{ }
}
