% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomize.R
\name{CheckISAEnrollmentStatus}
\alias{CheckISAEnrollmentStatus}
\title{CheckISAEnrollmentStatus}
\usage{
CheckISAEnrollmentStatus(vISAStatus, cRandomizer, dfCov)
}
\description{
{ This function will check the cRandomzier to make sure if an ISA is open for the covariate group dfCov}
}
