% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckNextTrialAnalysisTime.R
\name{CheckNextTrialAnalysisTime}
\alias{CheckNextTrialAnalysisTime}
\title{CheckNextTrialAnalysisTime}
\usage{
CheckNextTrialAnalysisTime(cISADesigns, lEnrolledPats, vISAStatus,
  dCurrentTime, vISAAnalysisIndx, vPreviousIATime)
}
\description{
{This function will Check the monitoring rule and return the time of next Analysis for the trial}
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/CheckNextTrialAnalysisTime.R}{View Code on GitHub} }
}
