% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTrialMonitor.R
\name{CheckTrialMonitor}
\alias{CheckTrialMonitor}
\title{Check Trial Monitor}
\usage{
CheckTrialMonitor(cISADesigns, lEnrolledPats, vISAStatus, dCurrentTime,
  vISAAnalysisIndx, vPreviousIATime)
}
\description{
{An ISA monitoring plan consists of the specifying a minimum number of patients and follow-up (FU) time
You must specify the minimum and the final analysis FU time.  This is done by specifying the
elements vMinQtyPats and  vMinFUTime  with at least 2 elements.   The first elements define when to
start monitoring based on a number of patients enrolled with additional FU and the last elements in the vectors
define when to do the final analysis .  For example, if an ISA will enroll up to 200 patients and  we define the elements in the following way
vMinQtyPats <- c( 30, 200 )
vMinFUTime  <- c( 3,  3)
would start monitoring patients 3 months after the 30th patient enrolled and the final analysis (FA) would be done 3 months after the 200th (last patient) enrolled

  To increase flexibility there are two options for adding additional IAs.  Which option is used is based on dQtyMonthsBtwIA ==0  or > 0
      Option 1:  the vMinQtyPats and vMinFUTime are of equal length with length > 2 then the trial would consist of length( vMinQtyPats )-1 IAs and 1 FA
          Example:    dQtyMonthsBtwIA  <-  0
                      vMinQtyPats     <- c( 30, 90, 150, 200 )
                      vMinFUTime      <- c( 3,  3,  0,    3)
              Would perform  IA 1 at 3 months after the 30th patient enrolled, IA 2 at 3 months after the 90th patient enrolled, IA 3 when the 150th patient enrolled,
              (not since the 3rd element of vMinFUTime = 0 IA 3 is performed when the 150th patient is enrolled) and the FA is performed 3 months after the 200th patient
              NOTE: With dQtyMonthsBtwIA > 0 and length( vMinQtyPats ) > 2 --> ERROR and stops running


      Option 2: You MUST have the lengths of vMinQtyPats and vMinFUTime equal to 2.
          Example:    dQtyMonthsBtwIA <-  2
                      vMinQtyPats     <- c( 30, 200 )
                      vMinFUTime      <- c( 3,   3)
                      This option will run the first IA 3 months after the 30th patient.  After that, the ISA is monitored every 2 months (vQtyMonthsBtwIA[1])
                      and the FA is 3 months after the 200th patient



       Return: list( vRunISAAnalysis, vPreviousIATime )
              vRunISAAnalysis = 1 if the ISA needs to have the IA run and 0 otherwise
              vPreviousIATime > 0 for the option when vQtyMonthsBtw[1] > 0 so we can track when the IA is done and know when the next one is to be done.

      In both cases, vIsFinalISAAnalysis = TRUE if dCurrentTime >= time of the FA as defined above; otherwise it is set to FALSE

      If vISAStatus != 1 then for the corresponding ISA elements in the return list this function returns
          vRunISAAnalysis = 0, vPreviousIATime = vPreviousIATime , vIsFinalISAAnalysis = FALSE, vCase = corresponding case )}
}
