% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCILimits.R
\name{GetCILimits}
\alias{GetCILimits}
\title{GetCILimits}
\usage{
GetCILimits(lAnalysis, nISAAnalysisIndx, bIsFinalISAAnalysis)
}
\description{
{ This function used to obtain the CI limits based on the analysis index and/or if it is the final analysis for an ISA. }
}
