% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeISARandomizer.R
\name{InitializeISARandomizer.EqualRandomizer}
\alias{InitializeISARandomizer.EqualRandomizer}
\title{InitializeISARandomizer.EqualRandomizer}
\usage{
\method{InitializeISARandomizer}{EqualRandomizer}(cISARandomizer, dISAStartTime)
}
\arguments{
\item{cISARandomizer}{Must contain elements cISARandomizer$vTrtLab and cISARandomizer$vQtyPats and the lengths must
be equal. If cISARandomizer contains a definition of cISARandomizer$dfSubgroupEnrollmentStatus it is used as the enrollment to a specific
covariate signature.}

\item{dISAStartTime}{Parameter is not utilized for EqualRandomizer}
}
\description{
{This function is used to initialize the EqualRandomzier for an ISA. It uses the vTrtLab and vQtyPats
to sample the desired number of each treatment in the ISA. }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/InitializeISARandomizer.R}{View Code on GitHub} }
}
