% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeTrialRandomizer.R
\name{InitializeTrialRandomizer.default}
\alias{InitializeTrialRandomizer.default}
\alias{InitializeTrialRandomizer}
\title{InitializeTrialRandomizer.default}
\usage{
InitializeTrialRandomizer(cTrialDesign)
}
\description{
{This function is used to initialize the trial randomizer. The default version each
ISA and calls the InitializeISARandomizer.  Since each ISA can have a specific randomizer  the default
for the TrialRandomizer is just to initialize each ISA randomizer.
}
}
