% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.HighDoseOnly}
\alias{MakeDecisionDoses.HighDoseOnly}
\title{MakeDecisionDoses.HighDoseOnly}
\usage{
\method{MakeDecisionDoses}{HighDoseOnly}(lDoseDec)
}
\description{
{"This option allows for multiple doses but only uses the high dose for decision making "  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/MakeDecisionDoses.R}{View Code on GitHub} }
}
