% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDecisionDoses.R
\name{MakeDecisionDoses.MakeDecisionMCPMod}
\alias{MakeDecisionDoses.MakeDecisionMCPMod}
\title{MakeDecisionDoses.MakeDecisionMCPMod}
\usage{
\method{MakeDecisionDoses}{MakeDecisionMCPMod}(lDoseDec)
}
\description{
MakeDecisionDoses.MakeDecisionMCPMod
}
