% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{ProcessData.default}
\alias{ProcessData.default}
\title{ProcessData.default.}
\usage{
\method{ProcessData}{default}(cDS, dCurrentTime)
}
\description{
{ No default version is defined to varify that the user correctly specifies the desired function.
This function is used to process the data before the RunAnalysis function is called.  This can
be useful for cases where the data is simulated/collected in one for but another form is needed for analysis.  For
example, the patient data over time is simualted but the data is analyzed as change from baseline. }
}
