% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.Ranked.R
\name{RunAnalysis.Ranked}
\alias{RunAnalysis.Ranked}
\title{RunAnalysis.Ranked}
\usage{
\method{RunAnalysis}{Ranked}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis, cRandomizer)
}
\description{
{This is the method used for running the Ranked analysis. Specifically, wilcox.test with exact = TRUE"  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunAnalysis.Ranked.R}{View Code on GitHub} }
}
