% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.RankedExact.R
\name{RunAnalysis.RankedExact}
\alias{RunAnalysis.RankedExact}
\title{RunAnalysis.RankedExact}
\usage{
\method{RunAnalysis}{RankedExact}(cAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis, cRandomizer)
}
\description{
{This is the method used for running the Ranked analysis using wilcox_test from the coin package.
Specifically, wilcox_test (DO NO CONFUSE with wilcox.test) with exact = TRUE"  }
}
