% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalysis.R
\name{RunAnalysis}
\alias{RunAnalysis}
\title{RunAnalysis}
\usage{
RunAnalysis(cAnalysis, lDataAna, nISAAnalysisIndx, bIsFinalISAAnalysis)
}
\description{
{This is the method used for running the analysis.  There are several options available for different types
of data.}
}
