% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFinalTrialAnalysis.R
\name{RunFinalAnalysis}
\alias{RunFinalAnalysis}
\alias{RunFinalTrialAnalysis}
\title{RunFinalAnalysis}
\usage{
RunFinalTrialAnalysis(cISADesigns, cEnrolledPats, vISAStatus,
  vISAAnalysisIndx)
}
\description{
{ Need description }
}
