% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunISAAnalysis.R
\name{RunISAAnalysis}
\alias{RunISAAnalysis}
\title{RunISAAnalysis}
\usage{
RunISAAnalysis(cISAAnalysis, lDataAna, nISAAnalysisIndx,
  bIsFinalISAAnalysis)
}
\description{
{Generic method called to run the ISA analysis.  }
}
