% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRealTrialAnalysis.R
\name{RunRealTrialAnalysis}
\alias{RunRealTrialAnalysis}
\title{RunRealTrialAnalysis}
\usage{
RunRealTrialAnalysis(cTrialDesign, strInputFile, strOutputFile, nSeed)
}
\description{
{Generic function that could be used to run an analysis on trial data.  }
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/RunRealTrialAnalysis.R}{View Code on GitHub} }
}
