% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSimulation.R
\name{RunSimulation.default}
\alias{RunSimulation.default}
\title{RunSimulation.RunSimulation}
\usage{
\method{RunSimulation}{default}(lSimulation)
}
\description{
{Generic function to execute the simulation of a list of scenarios contained in lScenarios.  }
}
