% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTrialAnalysis.R
\name{RunTrialAnalysis}
\alias{RunTrialAnalysis}
\alias{RunTrialAnalysis.default}
\title{RunTrialAnalysis}
\usage{
RunTrialAnalysis(cISADesigns, cEnrolledPats, vISAStatus, dCurrentTime,
  vRunISAAnalysis, vISAAnalysisIndx, vIsFinalISAAnalysis)

\method{RunTrialAnalysis}{default}(cISADesigns, cEnrolledPats, vISAStatus,
  dCurrentTime, vRunISAAnalysis, vISAAnalysisIndx, vIsFinalISAAnalysis)
}
\description{
{ RunTrialAnalysis is a generic method to execute the trial level analysis, eg execute all ISA analysis.
Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA   }

{ RunTrialAnalysis is a generic method to execute the trial analysis.
Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA.
Trial level analysis - will go through and conduct the analysis for each ISA.
 This version is sent all the available data in lDataAna then for each ISA gets the subset
  that is specific to that ISA.  If we want to do a version that borrows control info
  a new version could be adapted.
 }
}
