% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPatientOutcomes.MVNWithCovariates.R
\name{SimPatientOutcomes.MVNWithCovariates}
\alias{SimPatientOutcomes.MVNWithCovariates}
\alias{{SimPatientOutcomes.MVNWithCovariates}}
\title{SimPatientOutcomes.MVNWithCovariates}
\usage{
\method{SimPatientOutcomes}{MVNWithCovariates}(cSimOutcomes, cISADesign, dfPatCovISA)
}
\description{
In this SimPatientOutcomes
}
\seealso{
{ \href{https://github.com/kwathen/OCTOPUS/blob/master/R/SimPatientOutcomes.MVNWithCovariates.R}{View Code on GitHub} }
}
