% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccrualMethods.R
\docType{methods}
\name{SimulateAdditionalArrivalTimes}
\alias{SimulateAdditionalArrivalTimes}
\title{SimulateAdditionalArrivalTimes}
\usage{
\S4method{SimulateAdditionalArrivalTimes}{AccrualMethods}(cAP, nQtyPats,
  vCurrentStartTimes)
}
\arguments{
\item{nQtyPats}{Number of additonal times to simulate}

\item{vCurrentStartTimes}{a vector of the current start times.   The new times will be after the last element and the vector appended to.}
}
\description{
{
}
}
\details{
{ }
}
\examples{
  vPatsPerMonth  <- c(5, 10, 15, 20, 30, 40, 50) # The ramp up in expected Pat/month
  nMaxQtyPats    <- 100  # Maximum of 100 patients
  ap             <- NewAccrualProcess( vQtyPatsPerMonth = vPatsPerMonth, nMaxQtyPatients = nMaxQtyPats )
  vAccTimes      <- SimulateArrivalTimes( ap )

}
\seealso{
\code{\link{NewAccrualProcess}} for creating the object.
}
