% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePatientCovariates.Categorical.R,
%   R/SimulatePatientCovariates.R
\name{SimulatePatientCovariates.Categorical}
\alias{SimulatePatientCovariates.Categorical}
\alias{SimulatePatientCovariates}
\alias{SimulatePatientCovariates.default}
\title{SimulatePatientCovariates.Categorical}
\usage{
\method{SimulatePatientCovariates}{Categorical}(cSimCovariate,
  nQtyOfPatients)

SimulatePatientCovariates(cSimCovariate, nQtyOfPatients)

\method{SimulatePatientCovariates}{default}(cSimCovariate, nQtyOfPatients)
}
\description{
SimulatePatientCovariates {This function is used to simulate the patient covariate values.
This function is intended to do one covariate for all patients.  }
}
\section{Methods (by class)}{
\itemize{
\item \code{Categorical}: { This function is used to simulate the patient covariate values.
This function is intended to do one covariate, for all patients.  The sampled covaraites values will be 1,2,..., length( cSimCovariate$vProbs ).}

\item \code{default}: { This function is used to simulate the patient covariate values.
This function is intended to do one covariate, for all patients.
Because several options are provided and there in no well defined default
an stop error occurs if you call the default method. }
}}

