% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_database.R
\name{view_database}
\alias{view_database}
\title{View Database Connection with Octopus}
\usage{
view_database(con, options = list())
}
\arguments{
\item{con}{A database connection object. The result of DBI::dbConnect().}

\item{options}{A named list of options to be passed along to shinyApp().}
}
\value{
An R Shiny instance.
}
\description{
This function opens a shiny instance where the database can
  be viewed.
}
